% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clusterings.R
\name{add_clusterings}
\alias{add_clusterings}
\title{Adds a clustering to the clustering list}
\usage{
add_clusterings(clusterings = NULL, clusterings_to_add = NULL)
}
\arguments{
\item{clusterings}{clusterings object}

\item{clusterings_to_add}{data.frame of 1 or more cell labels with cell id as row names}
}
\description{
Adds a clustering to the clustering list
}
