% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_features}
\alias{add_features}
\title{Adds a feature to the feature list}
\usage{
add_features(scMuffinList = NULL, id = NULL, summary = NULL, full = NULL)
}
\arguments{
\item{features_obj}{features object}

\item{to_add}{data.frame of 1 or more cell values with cell id as row names}
}
\description{
Adds a feature to the feature list
}
