% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_partitions.R
\name{add_partitions}
\alias{add_partitions}
\title{Adds a clustering to the clustering list}
\usage{
add_partitions(scMuffinList = NULL, clusters = NULL, partition_id = NULL)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{clusters}{named factor where the values are cells names and names are cluster labels}

\item{partition_id}{unique identifier for this partition}
}
\value{
scMuffinList with the additional partition
}
\description{
Add this partition to the set of partitions
}
