% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_outliers_col.R
\name{adj_outliers_col}
\alias{adj_outliers_col}
\title{Remove outliers to avoid that these influence colors}
\usage{
adj_outliers_col(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
The definition of outliers is based on the function 'boxplot.stats()'. Values of x lower than the extreme of the lower whisker are set equal to it; values of x greater than the exteme of the upper whisker are set equal to it.
}
