% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_CNV_reference.R
\name{apply_CNV_reference}
\alias{apply_CNV_reference}
\title{apply_CNV_reference}
\usage{
apply_CNV_reference(
  cnv = NULL,
  cnv_clustering = NULL,
  reference = "reference",
  eps = 0.2,
  method = c("mean", "min_max")
)
}
\arguments{
\item{cnv}{the output of calculate_CNV}

\item{cnv_clustering}{the output of cluster_by_features(..., cnv=TRUE, ...)}

\item{reference}{name of the reference column}

\item{eps}{threshold for the min_max method}

\item{method}{"mean" (default): subtract the average profile of the reference cluster to every cell; "min_max" see Tirosh et al. (DOI: 10.1126/science.aad0501). In the "min_max" method all values below (above) the maximum (minimum) of the reference cluster + (-) eps are set to 0. The lower (upper) extreme whisker are considered as minimum (maximum) values.}
}
\description{
Add reference vector to CNV matrix
}
