% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_cluster_enrichment.R
\name{assess_cluster_enrichment}
\alias{assess_cluster_enrichment}
\title{Assess enrichment with ORA or CSEA}
\usage{
assess_cluster_enrichment(
  scMuffinList = NULL,
  feature_name = NULL,
  partition_id = NULL,
  min.cells.feature = 100,
  min.cells.cluster = 10,
  fract_min = 0.2,
  mc.cores = 1,
  csea.k = 99,
  min.k = 10
)
}
\arguments{
\item{scMuffinList}{scMuffinList list}

\item{feature_name}{the names of the feature that should be considered. It must be one of names(scMuffinList)}

\item{partition_id}{identifier of the partition to be considered}

\item{min.cells.feature}{minimum number of cells in which a feature must have value}

\item{min.cells.cluster}{minimum number of cells of a cluster}

\item{fract_min}{only cluster of size less or equal to this fraction of cell with not null feature values will be analysed}

\item{mc.cores}{number of cores}

\item{csea.k}{number of CSEA permutations}

\item{min.k}{minimum number of valid permutations to support empirical nulls}
}
\value{
scMuffinList with CSEA or ORA elements under scMuffinList$cluster_data for the considered partition
}
\description{
Assess cluster enrichment using ORA for categorical features and CSEA for numeric features.
}
\details{
The output of CSEA is a table with statistics for every tested gene set. 
The output of ORA is composed of a series of tables with enrichment results, one for every possible categorical value. See [extract_cluster_enrichment_table] to extract summary table from CSEA and ORA results.
}
