% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_cluster.R
\name{barplot_cluster}
\alias{barplot_cluster}
\title{Barplot clusters}
\usage{
barplot_cluster(
  scMuffinList = NULL,
  feature_id = NULL,
  feature_name = NULL,
  partition_id = NULL,
  dir_out = "./",
  n_features = 10,
  only_pos_nes = TRUE,
  do_scale_features = FALSE,
  cex.axis = 0.8,
  p.type = c("p", "p_adj")
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{feature_id}{names of the features to be used.}

\item{partition_id}{one among the partitions}

\item{dir_out}{string, output directory}

\item{n_features}{maximum number of features that will be shown}

\item{only_pos_nes}{whether to consider only positive enrichments}

\item{cex.axis}{cex.axis}

\item{quantity}{which column of the hypergeometric result. Default is "p". Other values are "wbd" and "er"}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\details{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
