% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster.R
\name{boxplot_cluster}
\alias{boxplot_cluster}
\title{Boxplot clusters}
\usage{
boxplot_cluster(
  features = NULL,
  cell_clusters = NULL,
  cluster_enrichment = NULL,
  dir_out = "./",
  only_top = 10,
  criterion = "fdr",
  fdr_threshold = NULL,
  only_pos_nes = TRUE,
  do_scale_features = FALSE,
  ...
)
}
\arguments{
\item{features}{feature list}

\item{cell_clusters}{cell clusters}

\item{cluster_enrichment}{result of assess_cluster_enrichment}

\item{dir_out}{output directory}

\item{only_top}{numeric, number of features to be shown with a different color and representing the most significative features according to t-test}

\item{criterion}{"fdr" to sort features by fdr}

\item{fdr_threshold}{fdr threshold}

\item{only_pos_nes}{whether to consider only positive enrichments}

\item{do_scale_features}{whether to scale features}

\item{...}{passed to axis}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
