% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster_.R
\name{boxplot_cluster_}
\alias{boxplot_cluster_}
\title{Boxplot clusters}
\usage{
boxplot_cluster_(
  features,
  cell_clusters,
  top_features = NULL,
  dir_out = "./",
  only_top = FALSE
)
}
\arguments{
\item{cell_clusters}{array with cell values named by their cluster ID}

\item{dir_out}{string, output directory}

\item{features_by_cells}{matrix, features by cells matrix}

\item{n_top}{numeric, number of features to be shown with a difference color and representing the most significative features according to t-test}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Noemi Di Nanni
}
