% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_points.R
\name{boxplot_points}
\alias{boxplot_points}
\title{Boxplot with points}
\usage{
boxplot_points(
  x = NULL,
  f = NULL,
  col = NULL,
  amount = 0.2,
  adj.col = 1,
  pch = 16,
  cex = 0.6,
  ylim = NULL,
  file = NULL,
  width = 180,
  height = 180,
  units = "mm",
  res = 300,
  image_format = "png",
  ...
)
}
\arguments{
\item{x}{values}

\item{f}{factor to split the values}

\item{col}{color palette. Equal in length to levels(f)}

\item{amount}{amount of jitter}

\item{adj.col}{color adjustment}

\item{pch}{point pch}

\item{cex}{cex}

\item{ylim}{ylim}

\item{file}{output file}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{image_format}{png or jpeg}

\item{...}{further argument to boxplot}
}
\description{
Boxplot with points
}
