% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cluster_by_gs.R
\name{calc_cluster_by_gs}
\alias{calc_cluster_by_gs}
\title{calc_cluster_by_gs}
\usage{
calc_cluster_by_gs(genes_by_cells, gene_sets, out_dir = "./")
}
\arguments{
\item{genes_by_cells}{matrix}

\item{gene_sets}{list of gene sets}

\item{out_dir}{output directory}
}
\description{
calc_cluster_by_gs
}
