% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CNV.R
\name{calculate_CNV}
\alias{calculate_CNV}
\title{Calculate CNV}
\usage{
calculate_CNV(
  genes_by_cells,
  reference = NULL,
  mc.cores = 2,
  wnd_size = 100,
  min_genes = 1000,
  min_cells = 100,
  scale_cells = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{reference}{NULL. It is possible to add a reference vector, downloaded from GTEx portal.}

\item{genes_by_cell}{Preprocessed Seurat object}
}
\description{
Calculate CNV
}
