% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gs_scores.R
\name{calculate_gs_scores}
\alias{calculate_gs_scores}
\title{Calculate gene set scores}
\usage{
calculate_gs_scores(
  genes_by_cells,
  gs_list = NULL,
  mc.cores = 2,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 10,
  k = 100,
  kmin = 50,
  score_type = c("relative", "mean"),
  null_model = TRUE,
  verbose = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{genes_by_cells}{genes_by_cells expression matrix. This matrix must contain positive values only.}

\item{gs_list}{list of gene sets}

\item{mc.cores}{number of cores}

\item{nbins}{number of bins to split the distribution of average gene expression}

\item{nmark_min}{numner of minimum markers that are required for the succesful calculation of a gene set score}

\item{ncells_min}{numner of minimum cells in which a gene set has to be succesfully calculated}

\item{k}{number of permutations}

\item{kmin}{minimum number of permutations; due to missing values it is hard to ensure that a gene set score can be compared to k permutations in every cell}

\item{score_type}{type of score. if "relative", than the score is the difference between the observed gene set average expression and that of a k permutations; if "mean" the score is equal to the observed gene set average expression}

\item{null_model}{TRUE if permutations have to be used. Required for score_type="relative"}

\item{verbose}{verbosity}

\item{na.rm}{whether to use NA or not}
}
\value{
list of objects containing also the feature object (feat_obj) for visualization purpose
}
\description{
Calculate gene set scoring
}
