% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gs_scores_in_clusters.R
\name{calculate_gs_scores_in_clusters}
\alias{calculate_gs_scores_in_clusters}
\title{Calculate gene set score at cluster level}
\usage{
calculate_gs_scores_in_clusters(
  scMuffinList = NULL,
  partition_id = NULL,
  ncells_min = 5,
  null_model = TRUE,
  alt = "g",
  test = "t"
)
}
\arguments{
\item{partition_id}{cell cluster labels}

\item{ncells_min}{min number of cells required for the clculation of the average signature in the cluster}

\item{null_model}{TRUE to consider the permutations}

\item{alt}{alterative}

\item{test}{type of test}

\item{gs_scores_obj}{list of gene signature lists}
}
\value{
scMuffinList with updates in elements cluster_data and cluster_data_full
}
\description{
Calculate gene set score at cluster level
}
