% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signatures.R
\name{calculate_signatures}
\alias{calculate_signatures}
\title{calculate signatures}
\usage{
calculate_signatures(
  genes_by_cells,
  signatures = NULL,
  mc.cores = 2,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50,
  score_type = c("relative", "mean"),
  null_model = TRUE
)
}
\description{
calculate signatures
}
