% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signatures_clusters.R
\name{calculate_signatures_clusters}
\alias{calculate_signatures_clusters}
\title{calculate_signatures_clusters}
\usage{
calculate_signatures_clusters(
  sign_list,
  cell_clusters = cell_clusters,
  ncells_min = 5,
  null_model = TRUE
)
}
\arguments{
\item{sign_list}{list of gene signature lists}

\item{cell_clusters}{cell cluster labels}

\item{ncells_min}{min number of cells required for the clculation of the average signature in the cluster}

\item{null_model}{TRUE to consider the permutations}
}
\description{
calculate_signatures_clusters
}
