% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_coocc_partitions.R
\name{cel_coocc_partitions}
\alias{cel_coocc_partitions}
\title{Co-occurrence of cells between two partitions}
\usage{
cel_coocc_partitions(p1 = NULL, p2 = NULL)
}
\arguments{
\item{p1}{a partition among scMuffinList$partitions, e.g. scML_demo$partitions$global_exp}

\item{p2}{a partition among scMuffinList$partitions, e.g. scML_demo$partitions$CNV}
}
\value{
matrix of Jaccard indeces between all levels of p1 and all levels of p2
}
\description{
Calculate, as Jaccard index, the co-occurrence of cells in all the pairs of elements composed by an element from p1 and an element from p2.
}
