% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_by_features.R
\name{cluster_by_features}
\alias{cluster_by_features}
\title{Cluster by features}
\usage{
cluster_by_features(features_by_cells = NULL, n_comp = 10, ...)
}
\arguments{
\item{features_by_cells}{features by cells}

\item{n_comp}{numeric, Dimensions of reduction to use as input}

\item{...}{arguments passed to Seurat::FindCLusters}
}
\value{
features_by_cells Seurat Object, object with saved dimension reduction components calculate on features by cells matrix
}
\description{
Clustering of the features_by_cell matrix
}
