% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_by_gs.R
\name{cluster_by_gs}
\alias{cluster_by_gs}
\title{cluster by n-genes}
\usage{
cluster_by_gs(genes_by_cells, gs = NULL)
}
\arguments{
\item{genes_by_cells}{Seurat object}

\item{gs}{vector of genes}
}
\description{
cluster by n-genes
}
