% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_comparison.R
\name{cluster_comparison}
\alias{cluster_comparison}
\title{cluster_comparison}
\usage{
cluster_comparison(cl1, cl2, ctree1, ctree2)
}
\arguments{
\item{cl1}{x}

\item{cl2}{x}

\item{ctree1}{x}

\item{ctree2}{x}
}
\description{
Cluster comparison
}
