% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_csea.R
\name{cluster_csea}
\alias{cluster_csea}
\title{Calculate cluster enrichment by means of CSEA}
\usage{
cluster_csea(
  feature_values = NULL,
  cell_clusters = NULL,
  min.cells.feature = 100,
  min.cells.cluster = 10,
  mc.cores = 1,
  csea.k = 99,
  min.k = 10,
  fract_min = 0.2
)
}
\arguments{
\item{feature_values}{matrix, features by cells matrix}

\item{cell_clusters}{array with cell values named by their cluster ID}

\item{min.cells.feature}{minimum number of cells in which a feature must have value}

\item{min.cells.cluster}{minimum number of cells of a cluster}

\item{mc.cores}{number of cores}

\item{csea.k}{number of permutations}

\item{min.k}{minimum number of valid permutations to support empirical nulls}

\item{fract_min}{only cluster of size less or equal to this fraction of cell with not null feature values will be analysed}
}
\value{
list with two elements: gs_table and leading_edge. See [csea()]
}
\description{
Calculate cluster enrichment by csea approach
}
