% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrichment.R
\name{cluster_enrichment}
\alias{cluster_enrichment}
\title{Calculate cluster enrichment}
\usage{
cluster_enrichment(features_by_cells, cell_clusters, n_top = 3, dir_out = "./")
}
\arguments{
\item{features_by_cells}{matrix, features by cells matrix}

\item{cell_clusters}{array with cell values named by their cluster ID}

\item{n_top}{numeric, number of features to be shown with a difference color and representing the most significative features according to t-test}

\item{dir_out}{string, output directory}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Ettore Mosca
}
