% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hyper.R
\name{cluster_hyper}
\alias{cluster_hyper}
\title{Calculate cluster enrichment by hypergeometric test}
\usage{
cluster_hyper(feature_values = NULL, cell_clusters = NULL)
}
\arguments{
\item{feature_values}{matrix, cells-by-features}

\item{cell_clusters}{array with cell values named by their cluster ID}
}
\value{
A list of data.frames, one for each categorical value. See [ora()]
}
\description{
Calculate cluster enrichment in each of the categorical values of a feature, by means of hypergeometric test
}
\author{
Ettore Mosca
}
