% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings_to_list.R
\name{clusterings_to_list}
\alias{clusterings_to_list}
\title{clusterings_to_list}
\usage{
clusterings_to_list(clusterings)
}
\arguments{
\item{clusterings}{clusterings object}
}
\description{
clusterings_to_list
}
