% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{CNV}
\alias{CNV}
\title{CNV}
\usage{
result_cnv <- mclapply(mat_sorted, function(x) apply(x[,5:dim(x)[2]], 2, function( y ) CNV(y)), mc.cores = 2)
}
\arguments{
\item{x}{list of dataframes retrieved by 'preprocess_object_for_cnv'.}
}
\description{
Function to compute CNV.
}
\details{
As a list of dataframes, it is convenient to use 'mclapply' from parallel package.
}
\references{
Patel paper supplementary material (?)
}
\author{
Valentina Nale
}
