% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_analysis.R
\name{CNV_analysis}
\alias{CNV_analysis}
\title{CNV analysis}
\usage{
CNV_analysis(
  scMuffinList = NULL,
  mc.cores = 1,
  reference = NULL,
  min_cells = 100,
  min_genes = 200,
  wnd_size = 100,
  center_cells = TRUE,
  center_genes = FALSE,
  expr_lim = FALSE,
  method = "mean",
  na.rm = FALSE,
  z.score = FALSE,
  eps = NULL,
  n_comp = 10,
  gene_ann = NULL,
  ...
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{mc.cores}{number of cores;}

\item{reference}{one-column data.frame with a reference expression profile; rownames must match those of genes_by_cell;}

\item{min_cells}{minimum numbver of cells in which a gene must be expressed;}

\item{min_genes}{minimun number of genes expressed in a cell;}

\item{wnd_size}{number of adjacent genes considered;}

\item{center_cells}{whether to center cells. Default is TRUE}

\item{center_genes}{whether to center genes or not. Default is FALSE. When using a reference it is useful to set this to TRUE}

\item{expr_lim}{min and max values of relative expression; by default, lower and higher whiskers returned by grDevices::boxplot.stats will be used. Set to NA or anything other value such that length(expr_lim) != 2 to disable the removal of outliers.}

\item{method}{"mean": subtract the average profile of the reference cluster to every cell (default); "min_max" (from Tirosh et al., DOI: 10.1126/science.aad0501) can be used but is not tested yet}

\item{na.rm}{whether to remove 0 values in CNV estimation}

\item{z.score}{whether to use z-scores of the cluster median CNV profile, instead of the median itself.}

\item{eps}{absolute threshold to call CNV regions.}

\item{n_comp}{number of PCA components used for clustering}

\item{gene_ann}{optional data.frame with gene annotation with mandatory columns "Chromosome", "symbol" and "pos" (genomic location). If NULL gene locations will be collected from org.Hs.eg.db}

\item{...}{arguments passed to Seurat::FindClusters()}
}
\description{
CNV analysis
}
