% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusterings.R
\name{create_clusterings}
\alias{create_clusterings}
\title{Create clustering list}
\usage{
create_clusterings(cell_id = NULL, values = NULL)
}
\arguments{
\item{cell_id}{vector of cell id}

\item{values}{vector of values}
}
\description{
Create clustering list
}
