% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_features.R
\name{create_features}
\alias{create_features}
\title{Create feature list}
\usage{
create_features(x = NULL)
}
\arguments{
\item{x}{data.frame of 1 or more cell values with cell id as row names}
}
\description{
Create feature list
}
