% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_partitions_obj.R
\name{create_partitions_obj}
\alias{create_partitions_obj}
\title{Create the clustering object}
\usage{
create_partitions_obj(x)
}
\arguments{
\item{x}{data.frame of 1 or more cell labels with cell id as row names}
}
\description{
Create the clustering object
}
