% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scMuffinList.R
\name{create_scMuffinList}
\alias{create_scMuffinList}
\title{Create scMuffinList}
\usage{
create_scMuffinList(counts = NULL, normalized = NULL)
}
\arguments{
\item{counts}{genes-by-cells matrix or data.frame of counts. Rownames must be gene identifiers.}

\item{normalized}{genes-by-cells matrix or data.frame with normalized expression. Rownames must be gene identifiers.}
}
\value{
scMuffinList with counts and normalized elements as dgCMatrix.
}
\description{
Create scMuffinList from counts and/or normalized data. The gene identifiers must be the same. Note that CNV analysis is currently available only for symbols as gene identifiers.
}
