% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_cluster.R
\name{dotplot_cluster}
\alias{dotplot_cluster}
\title{Boxplot clusters}
\usage{
dotplot_cluster(
  cells_by_features,
  cell_clusters,
  top_features = NULL,
  dir_out = "./",
  cont_tables = NULL
)
}
\arguments{
\item{cells_by_features}{matrix, features by cells matrix}

\item{cell_clusters}{array with cell values named by their cluster ID}

\item{dir_out}{string, output directory}

\item{n_top}{numeric, number of features to be shown with a difference color and representing the most significative features according to t-test}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Noemi Di Nanni
}
