% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{es}
\alias{es}
\title{Enrichment Score}
\usage{
es(idx, x, le = F)
}
\arguments{
\item{idx}{vector of indices a subset of elements of x}

\item{x}{named vector, ranked list}

\item{le}{leading edge information}
}
\value{
enrichment score
}
\description{
Enrichment Score
}
