% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_rate.R
\name{exp_rate}
\alias{exp_rate}
\title{Expression_rate}
\usage{
exp_rate(genes_by_cells, ngenes_min = 5)
}
\arguments{
\item{genes_by_cells}{matrix, real data}

\item{ngenes_min}{min numeric, number of genes in each cell}
}
\value{
expRate_vect vector with expression rate for each cell
}
\description{
Ratio between the number of expressed genes and the total number of reads
}
\author{
Noemi Di Nanni
}
