% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_table.R
\name{extract_cluster_enrichment_table}
\alias{extract_cluster_enrichment_table}
\title{Extract cluster enrichment table}
\usage{
extract_cluster_enrichment_table(
  scMuffinList,
  partition_id = NULL,
  type = NULL,
  quantity = NULL,
  feature_name = NULL
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{partition_id}{id of the parition to be used}

\item{type}{"GSEA" or "ORA"}

\item{quantity}{for GSEA: "es", "p_val", "adj_p_val", "nes" or "FDRq". ORA: "wbd", "exp", "er", "p", "p_adj".}

\item{feature_name}{name of the features that will be considered. Must be any of `colnames(scMuffinList$cluster_data$[[parition_id]]$GSEA$gs_table)` or `names(scMuffinList$cluster_data$[[parition_id]]$ORA)`}
}
\description{
Extract cluster enrichment table
}
