% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_tags.R
\name{extract_cluster_enrichment_tags}
\alias{extract_cluster_enrichment_tags}
\title{Extract the top results of CSEA and ORA for each cluster}
\usage{
extract_cluster_enrichment_tags(
  scMuffinList = NULL,
  partition_id = NULL,
  CSEA_selection_criterion = "FDRq",
  CSEA_selection_threshold = 0.25,
  only_pos_nes = TRUE,
  ORA_selection_criterion = "p_adj",
  ORA_selection_threshold = 0.1,
  n_max_per_cluster = 3
)
}
\arguments{
\item{scMuffinList}{scMuffinList object.}

\item{partition_id}{one among the partitions.}

\item{CSEA_selection_criterion}{selection criteria for results of CSEA. See \code{\link{extract_cluster_enrichment_table}} for possible values.}

\item{CSEA_selection_threshold}{threshold for selection of tags from CSEA.}

\item{only_pos_nes}{whether only positive nes should be considered in CSEA.}

\item{ORA_selection_criterion}{selection criteria for results of ORA. See \code{\link{extract_cluster_enrichment_table}} for possible values.}

\item{ORA_selection_threshold}{threshold for selection of tags from ORA.}

\item{n_max_per_cluster}{maximum number of tags per cluster.}
}
\value{
Add or overwrite the "cluster_tags" list to `scMuffinList$cluster_data[[partition_id]]`
}
\description{
Extract the best results of CSEA and ORA for each cluster.
}
