% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gsl.R
\name{filter_gsl}
\alias{filter_gsl}
\title{Filter a gene set list}
\usage{
filter_gsl(gsl, universe, min_size = 5, max_size = 500)
}
\arguments{
\item{gsl}{gene set list (named list)}

\item{universe}{set of all possible values for items of gsl}

\item{min_size}{minimum gene set size}

\item{max_size}{maximum gene set size}
}
\value{
Filtered gene set list
}
\description{
Filter a gene set list
}
