% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_clusters.R
\name{find_clusters}
\alias{find_clusters}
\title{find_clusters}
\usage{
find_clusters(seurat_object, dims = 1:10)
}
\arguments{
\item{seurat_object}{seurat object}

\item{dims}{number of dimensions}
}
\description{
find_clusters
}
