% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_score_in_clusters.R
\name{gene_set_score_in_clusters}
\alias{gene_set_score_in_clusters}
\title{gene_set_score_in_clusters}
\usage{
gene_set_score_in_clusters(
  current_gene_set,
  seurat_object_data,
  seurat_object_ident,
  nmark_min = 5,
  perc = 0.75,
  ncells_min = NULL,
  bins = NULL,
  k = 100,
  alt = "g",
  test = "t"
)
}
\arguments{
\item{current_gene_set}{real data}

\item{seurat_object_data}{control data}

\item{seurat_object_ident}{min number of markers}

\item{nmark_min}{min number of cells}

\item{perc}{min percentage of gene set}

\item{ncells_min}{min number of cells}

\item{bins}{data bins}

\item{k}{number of permutations}

\item{alt}{alterative}

\item{test}{type of test}
}
\description{
gene_set_score_in_clusters
}
