% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_by_cells.R
\docType{data}
\name{genes_by_cells}
\alias{genes_by_cells}
\title{Seurat object, example dataset}
\format{
An object of class \code{Seurat} with 27298 rows and 3085 columns.
}
\usage{
genes_by_cells
}
\description{
From /home/bioinformatics/ndinanni/SC/SC_2/Yuan2018/PJ016/data_PJ016.RData
}
\keyword{datasets}
