% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_in_detected_CNV_regions.R
\name{genes_in_detected_CNV_regions}
\alias{genes_in_detected_CNV_regions}
\title{Retrieve genes in detected CNV regions}
\usage{
genes_in_detected_CNV_regions(scMuffinList = NULL)
}
\arguments{
\item{scMuffinList}{scMuffinList object}
}
\description{
Retrieve genes in detected CNV regions
}
