% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_clusters.R
\name{get_meta_clusters}
\alias{get_meta_clusters}
\title{Get cells in each meta cluster}
\usage{
get_meta_clusters(cl_list, meta_clusters)
}
\arguments{
\item{cl_list}{cluterings list}

\item{meta_clusters}{meta clusters}
}
\description{
Get cells in each meta cluster
}
