% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_msigdb_geneset.R
\name{get_msigdb_geneset}
\alias{get_msigdb_geneset}
\title{Returns msigdb geneset in a format compatible with calculate_signatures()}
\usage{
get_msigdb_geneset(
  species,
  category = NULL,
  subcategory = NULL,
  type = "gene_symbol"
)
}
\arguments{
\item{species}{Species name}

\item{category}{MSigDB category name, see msigdbr_collections()}

\item{subcategory}{MsigDB subcategory name, see msigdbr_collections()}

\item{type}{Gene name of interest, can be gene_symbol or entrez_gene}
}
\description{
Returns msigdb geneset in a format compatible with calculate_signatures()
}
