% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{gsea}
\alias{gsea}
\title{Gene Set Enrichment Analysis}
\usage{
gsea(rl, gsl, k = 100, ord.mode = -1, mc_cores_path = 1, mc_cores_perm = 1)
}
\arguments{
\item{rl}{numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory}

\item{gsl}{named list of gene sets}

\item{k}{integer, number of permutations}

\item{ord.mode}{ordering mode: -1 -> descending; 1 ascending; must be of lenght equal to ncol(rl)}

\item{mc_cores_path}{number of cores to use for parallel calculation of gene set lists; the total number of cpu used will be mc_cores_path x mc_cores_perm}

\item{mc_cores_perm}{number of cores to use for parallel calculation of ranked list permutations; the total number of cpu used will be mc_cores_path x mc_cores_perm}
}
\value{
data.frame with es, nes, p-value, adjusted p-value and FDR q-value
}
\description{
Gene set enrichment analysis
}
