% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsls_Symbol.R
\docType{data}
\name{gsls_Symbol}
\alias{gsls_Symbol}
\title{Gene set lists (Symbol)
Gene set lists collected from the following sources:
- CancerSEA: http://biocc.hrbmu.edu.cn/CancerSEA/goDownload
- CM_cancer: http://bio-bigdata.hrbmu.edu.cn/CellMarker
- CM_normal: http://bio-bigdata.hrbmu.edu.cn/CellMarker
- PNDB: https://panglaodb.se
- Tirosh2016: 10.1126/science.aad0501
- Ribosomal proteins:}
\format{
An object of class \code{list} of length 6.
}
\usage{
gsls_Symbol
}
\description{
Gene set lists (Symbol)
Gene set lists collected from the following sources:
- CancerSEA: http://biocc.hrbmu.edu.cn/CancerSEA/goDownload
- CM_cancer: http://bio-bigdata.hrbmu.edu.cn/CellMarker
- CM_normal: http://bio-bigdata.hrbmu.edu.cn/CellMarker
- PNDB: https://panglaodb.se
- Tirosh2016: 10.1126/science.aad0501
- Ribosomal proteins:
}
\keyword{datasets}
