% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_cluster_enrichment.R
\name{heatmap_cluster_enrichment}
\alias{heatmap_cluster_enrichment}
\title{heatmap of signatures}
\usage{
heatmap_CNV(chr_merged)
}
\arguments{
\item{X}{...}

\item{file}{...}
}
\description{
Plot as an heatmap the resulting data from CNV.
}
\details{
Preprocessing with 'preprocess_for_heatmap' needed.
}
\author{
Ettore Mosca
}
