% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_CNV.R
\name{heatmap_CNV}
\alias{heatmap_CNV}
\title{CNV Heatmap}
\usage{
heatmap_CNV(
  scMuffinList = NULL,
  genes = NULL,
  genes.labels = FALSE,
  mark.detected.cnv = FALSE,
  file = NULL,
  width = 180,
  height = 180,
  units = "mm",
  res = 300,
  cluster_fontsize = 8,
  chrom_fontsize = 8,
  legend_fontsize = 8,
  genes.labels.fontsize = 8,
  ...
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{genes}{gene labels to show as row annotation}

\item{genes.labels}{whether to show gene labels or just mark their position}

\item{mark.detected.cnv}{whether to show detected CNV regions}

\item{file}{File name to save the figure as png file}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{cluster_fontsize}{cluster fontsize}

\item{chrom_fontsize}{chromosome fontsize}

\item{legend_fontsize}{legend fontsize}

\item{genes.labels.fontsize}{gene labels fontsize}

\item{...}{arguments passed to ComplexHeatmap::Heatmap}
}
\description{
CNV Heatmap
}
