% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_features_corr.R
\name{heatmap_features_corr}
\alias{heatmap_features_corr}
\title{heatmap_features_corr}
\usage{
heatmap_features_corr(
  X,
  file = "heatmap_features_corr.jpg",
  pal = NULL,
  n_colors = 11,
  width = 180,
  height = 180,
  res = 300,
  cex.axis = 0.6
)
}
\description{
heatmap_features_corr
}
