% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_signatures.R
\name{heatmap_signatures}
\alias{heatmap_signatures}
\title{heatmap of signatures}
\usage{
heatmap_CNV(chr_merged)
}
\arguments{
\item{chr_merged}{}
}
\description{
Plot as an heatmap the resulting data from CNV.
}
\details{
Preprocessing with 'preprocess_for_heatmap' needed.
}
\author{
Ettore Mosca
}
