% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_strongest_representative.R
\name{keep_strongest_representative}
\alias{keep_strongest_representative}
\title{Keep the feature with the average highest values}
\usage{
keep_strongest_representative(genes_by_cells = NULL, row.names = NULL)
}
\arguments{
\item{genes_by_cells}{genes-by-cells expression matrix}

\item{row.names}{row.names that have to be assigned}
}
\description{
Keep the feature with the average highest values
}
