% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landSCENT.R
\name{landSCENT}
\alias{landSCENT}
\title{LandSCENT}
\usage{
landSCENT(genes_by_cells, ppi, reduceMethod = "PCA", clusterMethod = "dbscan")
}
\arguments{
\item{genes_by_cells}{real data}

\item{ppi}{The adjacency matrix of a user-given PPI network with rownames and colnames labeling genes (same gene identifier as in genes_by_cells)}

\item{reduceMethod}{indicates the method to do dimension reduction: "PCA" or "tSNE"}

\item{clusterMethod}{indicates the method to do clustering: "dbscan" or "PAM"}
}
\value{
landscent_list list with the following elements: SR, DPT, potency_states, complete_output
}
\description{
LandSCENT
}
