% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusterings_enrichment.R
\name{merge_clusterings_enrichment}
\alias{merge_clusterings_enrichment}
\title{merge_clusterings_enrichment}
\usage{
merge_clusterings_enrichment(clust_enrich)
}
\arguments{
\item{clust_enrich}{list of cluster enrichments}
}
\description{
merge_clusterings_enrichment
}
