% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_matrix.R
\name{merge_matrix}
\alias{merge_matrix}
\title{merge_matrix}
\usage{
merge_matrix(
  signatures_by_cells = NULL,
  expr_score = NULL,
  output_landscent = NULL
)
}
\arguments{
\item{expr_score}{expression scores (output of exp_Rate)}

\item{features_by_cells}{features-by-cells input matrix}

\item{landscent_list}{list with the following elements: SR, DPT, potency_states, complete_output (output of landSCENT)}
}
\value{
matrix_Scaled matrix merged and scaled
}
\description{
merge_matrix
}
