% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{ora}
\alias{ora}
\title{Over Representation Analysis}
\usage{
ora(wb, bb, gsl, p_adj_method = "fdr")
}
\arguments{
\item{wb}{hits (white balls)}

\item{bb}{other elements (black balls)}

\item{gsl}{named list of sets}

\item{p_adj_method}{p value adjustment method, see p.adjust.methods}
}
\description{
Over representation analysis
}
