% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora1gs.R
\name{ora1gs}
\alias{ora1gs}
\title{Hypergeometric test on 1 dataset}
\usage{
ora1gs(wb, bb, bd)
}
\arguments{
\item{wb}{white balls}

\item{bb}{black balls}

\item{bd}{balls drawn}
}
\description{
Hypergeometric test on 1 dataset
}
