% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora_pipeline.R
\name{ora_pipeline}
\alias{ora_pipeline}
\title{ora pipeline}
\usage{
ora_pipeline(
  deg_list = NULL,
  universe = NULL,
  gs = NULL,
  gsid2name = NULL,
  mc.cores = 1,
  eg2sym = NULL,
  min_size = 5,
  max_size = 500,
  out_dir = "./",
  write_tables = FALSE,
  id = "ora"
)
}
\arguments{
\item{deg_list}{gene list of interest}

\item{universe}{all genes}

\item{gs}{list of gene lists}

\item{gsid2name}{data.frame for annotation of gene sets. It must contain the column "gsid" with gene set ids that match those in gs}

\item{mc.cores}{number of cores}

\item{eg2sym}{data.frame for translating gene ids into symbol. It must contain the columncs gene_id and symbol}

\item{min_size}{minimum gene set size}

\item{max_size}{maximum gene set size}

\item{out_dir}{output directory}

\item{write_tables}{whether to write or not the output}

\item{id}{project name}
}
\description{
Over representation analysis pipeline
}
