% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_coefficient.R
\name{overlap_coefficient}
\alias{overlap_coefficient}
\title{Calculate the overlap coefficient between two sets}
\usage{
overlap_coefficient(x, y)
}
\arguments{
\item{x}{vector of items}

\item{y}{vector of items}
}
\description{
Calculate the overlap coefficient between two sets
}
