% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_comparison.R
\name{plot_cluster_comparison}
\alias{plot_cluster_comparison}
\title{plot_cluster_comparison}
\usage{
plot_cluster_comparison(
  X,
  dend_row,
  dend_col,
  file = "heatmap_cluster_comparison.jpg",
  pal = NULL,
  n_colors = 9,
  width = 180,
  height = 180,
  res = 300,
  cex.axis = 0.6
)
}
\description{
plot_cluster_comparison
}
