% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_map.R
\name{plot_diff_map}
\alias{plot_diff_map}
\title{Plot diffusion map colored by other features}
\usage{
plot_diff_map(
  scMuffinList = NULL,
  columns = c(1:2),
  col_data = NULL,
  file = NULL,
  width = 200,
  height = 200,
  units = "mm",
  res = 300,
  adj_outliers = TRUE,
  scale_feature = FALSE,
  min_cells = 50,
  ...
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{columns}{columns of scMuffinList$diffusion_map_pseudo_t$summary to be used for x and y}

\item{col_data}{named vector with values that will be translated into colors. Names are cell names.}

\item{file}{File name to save the figure as png file}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{adj_outliers}{logical, whether to adjust the group.by scores, removing outliers}

\item{scale_feature}{logical, whether to scale col_data}

\item{min_cells}{minimum number of cells in which the feature must have a non-zero value}

\item{...}{further arguments to plot()}
}
\description{
Produce a scatter plot where cells are placed according to the results of diffusion map analysis and colored by values given in col_data.
}
