% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_cluster_colored_features.R
\name{plot_gene_cluster_colored_features}
\alias{plot_gene_cluster_colored_features}
\title{plot_gene_cluster_colored_features}
\usage{
plot_gene_cluster_colored_features(
  seurat_object,
  features_by_cells,
  n_colors = 11
)
}
\description{
plot_gene_cluster_colored_features
}
