% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meta_clusters.R
\name{plot_meta_clusters}
\alias{plot_meta_clusters}
\title{Plot meta clusters}
\usage{
plot_meta_clusters(
  ov_mat = NULL,
  meta_clusters = NULL,
  out_dir = "./",
  file = "/meta_clusters.jpg",
  ...
)
}
\description{
Plot meta clusters
}
