% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile_CNV.R
\name{plot_profile_CNV}
\alias{plot_profile_CNV}
\title{CNV Profile}
\usage{
plot_profile_CNV(
  scMuffinList = NULL,
  cluster = 0,
  z.score = TRUE,
  file = NULL,
  width = 300,
  height = 90,
  units = "mm",
  res = 300,
  image_format = "png",
  cex.points = 0.7,
  cex.lab = 0.7,
  cex.axis = 0.7,
  cex.main = 0.7
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{z.score}{whether to plot clustere median z-scores of CNV signal (TRUE) or cluster median CNV signal (FALSE).}

\item{file}{output file}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{image_format}{png or jpeg}
}
\description{
Plot an heatmap of the CNV.
}
\details{
CNV Profile of every cluster
}
