% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{plot_umap}
\usage{
plot_umap(
  seurat_object,
  file = "umap.jpg",
  color_by = "ident",
  pal = NULL,
  labels = NULL
)
}
\arguments{
\item{seurat_object}{seurat object, object with saved dimension reduction components}

\item{file}{string, file name output}

\item{color_by}{string, specification of a feature to colour by (e.g. cluster ID)}
}
\description{
Generate a UMAP visualization
}
