% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_colored_features.R
\name{plot_umap_colored_features}
\alias{plot_umap_colored_features}
\title{plot_umap_colored_features}
\usage{
plot_umap_colored_features(
  seurat_object,
  features,
  dir = "./",
  scale_feature = TRUE,
  feature_breaks = NULL,
  adj_outliers = FALSE,
  neg_eps = -0.1,
  pos_eps = 0.1,
  min_cells = 10,
  ...
)
}
\arguments{
\item{seurat_object}{seurat object, object with saved dimension reduction components calculate on genes by cells matrix}

\item{features}{matrix, features by cells matrix}

\item{adj_outliers}{logical, whether to adjust the feature scores, removing outliers}

\item{neg_eps}{small negative value to define the color for null values around zero, default to -0.1}

\item{pos_eps}{small positive value to define the color for null values around zero, default to =0.1}

\item{n_colors}{numeric, number of colors to be used}
}
\description{
Generate a UMAP visualization colored by feature distributions
}
