% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_expr_features.R
\name{plot_umap_expr_features}
\alias{plot_umap_expr_features}
\title{plot_umap_expr_features}
\usage{
plot_umap_expr_features(seurat_object, cells_by_features, dir = "./")
}
\arguments{
\item{seurat_object}{seurat object, object with saved dimension reduction components calculate on genes by cells matrix}

\item{cells_by_features}{matrix, cells by features}
}
\description{
Generate a UMAP visualization colored by feature distributions
}
