% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\docType{data}
\name{ppi}
\alias{ppi}
\title{ppi
This human protein-protein interaction network is derived from LandSCENT net17Jan16.m, which was collected Pathway Commons (www.pathwaycommons.org) (version Jan. 2016).}
\format{
An object of class \code{matrix} with 11751 rows and 11751 columns.
}
\usage{
ppi
}
\description{
ppi
This human protein-protein interaction network is derived from LandSCENT net17Jan16.m, which was collected Pathway Commons (www.pathwaycommons.org) (version Jan. 2016).
}
\keyword{datasets}
