% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cluster_markers_list.R
\name{prepare_cluster_markers_list}
\alias{prepare_cluster_markers_list}
\title{prepare_cluster_markers_list}
\usage{
prepare_cluster_markers_list(
  marker_table_list = NULL,
  effect_size = "avg_logFC",
  effect_size_thr = 0.25,
  sig_column = "p_val_adj",
  sig_thr = 0.1,
  id_column = 0,
  cluster = "cluster",
  top = 500
)
}
\description{
prepare_cluster_markers_list
}
