% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gsls.R
\name{prepare_gsls}
\alias{prepare_gsls}
\title{Prepare the gene set lists collected from various sources}
\usage{
prepare_gsls(
  gs_sources = NULL,
  custom_gsls = NULL,
  CM_tissues = NULL,
  PNDB_tissues = NULL,
  msigdb_hs_cat_subcat = NULL,
  genes_min = 5,
  genes_max = 500,
  id_type = c("Symbol", "EntrezID"),
  scMuffinList = NULL,
  genes = NULL
)
}
\arguments{
\item{gs_sources}{character vector, possible values: "CM_cancer", "CM_normal", "CancerSEA", "PNDB" and "msigdb"}

\item{custom_gsls}{named list of gene sets}

\item{CM_tissues}{character vector with strings to filter CellMArker db by tissue name}

\item{PNDB_tissues}{character vector with strings to filter Panglao db by tissue name}

\item{msigdb_hs_cat_subcat}{data.frame with species, category and subcategory: e.g. "Homo sapiens", "NA", "CP:KEGG"}

\item{genes_min}{minimum number of genes required in a gene set}

\item{genes_max}{maximum number of genes required in a gene set}

\item{scMuffinList}{scMuffinList object}
}
\value{
A lists of gene set lists
}
\description{
Prepare the gene set lists collected from various sources
}
