% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_GTEx_gene_reads.R
\name{process_GTEx_gene_reads}
\alias{process_GTEx_gene_reads}
\title{Process GTEx expression data}
\usage{
process_GTEx_gene_reads(
  geneReads = NULL,
  GTEx_annot = NULL,
  tissue = NULL,
  id_type = c("gene_id", "symbol"),
  nrows = -1,
  norm.type = "LogNormalize",
  ...
)
}
\arguments{
\item{geneReads}{GTEx gene reads "gct" file, e.g. "GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_reads.gct"}

\item{GTEx_annot}{GTEx Sample annotation file, e.g. "GTEx_Analysis_v8_Annotations_SampleAttributesDS.txt"}

\item{tissue}{tissue name, see param "SMTS" in annotation}

\item{id_type}{whether the output should be specified as gene_id or symbol}

\item{nrows}{limit the read to the top nrows. Useful for testing.}

\item{norm.type}{Method for normalization. See Seurat::NormalizeData.}

\item{...}{further arguments passed to functin Seurat::NormalizeData.}
}
\value{
average normalized gene expression values
}
\description{
GTEx gene reads file is processed to obtain the average normalized gene expression in a tissue
}
