% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_analysis.R
\name{proliferation_analysis}
\alias{proliferation_analysis}
\title{cell cycle - merge the two Tirosh signatures for proliferation analysis}
\usage{
proliferation_analysis(
  genes_by_cells,
  mc.cores = 2,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50,
  score_type = c("relative", "mean"),
  null_model = TRUE
)
}
\description{
cell cycle - merge the two Tirosh signatures for proliferation analysis
}
