% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_samples_similarity.R
\name{quantify_samples_similarity}
\alias{quantify_samples_similarity}
\title{Compare two samples on the basis of cluster markers}
\usage{
quantify_samples_similarity(
  gbc_1,
  gbc_2,
  clusters_1,
  clusters_2,
  cluster_markers_1,
  cluster_markers_2,
  genes_min = 3,
  genes_max = 500,
  mc.cores = 2,
  null_model = TRUE,
  ncells_min = 5,
  do_plot = TRUE,
  sample_name_1 = "S1",
  sample_name_2 = "S2",
  outfile = "cluster_similarity.jpg",
  pal = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  top_genes = FALSE,
  ...
)
}
\arguments{
\item{gbc_1}{genes-by-cells expression matrix of sample 1}

\item{gbc_2}{genes-by-cells expression matrix of sample 2}

\item{clusters_1}{named vector with cell clustering fo sample 1}

\item{clusters_2}{named vector with cell clustering fo sample 2}

\item{cluster_markers_1}{list of markers for each cluster of sample 1}

\item{cluster_markers_2}{list of markers for each cluster of sample 2}

\item{genes_min}{minimum number of genes required among the markers of a cluster}

\item{genes_max}{maximum number of genes required among the markers of a cluster}

\item{mc.cores}{number of cores}

\item{null_model}{whether to use or not the empirical null model. See calculate_signature}

\item{ncells_min}{minim number of cells in a cluster}

\item{do_plot}{whether to plot or not the results}

\item{sample_name_1}{sample 1 name}

\item{sample_name_2}{sample 2 name}

\item{outfile}{out file name}

\item{pal}{palette for the output heatmap}

\item{cluster_rows}{whether to cluster or not the rows}

\item{cluster_columns}{whether to cluster or not the columns}

\item{top_genes}{If specified, only the first top_genes genes of every element of the lists cluster_markers_1 and cluster_markers_2 will be considered. This implies that the markers for each clusters are considered sorted by decreasing relevance. Default is FALSE, which}

\item{...}{arguments passed to calculate_signatures}
}
\description{
Compare two samples on the basis of cluster markers
}
