% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_clustering.R
\name{re_clustering}
\alias{re_clustering}
\title{Clustering of the features_by_cell matrix}
\usage{
re_clustering(features_by_cells, n_comp = 10)
}
\arguments{
\item{n_comp}{Dimensions of reduction to use as input}

\item{matrix}{matrix of features}
}
\value{
features_by_cells Seurat Object
}
\description{
Clustering of the features_by_cell matrix
}
