% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_clustering_draft.R
\name{re_clustering_draft}
\alias{re_clustering_draft}
\title{gene_set_score_in_clusters}
\usage{
re_clustering_draft(
  features_by_cells,
  n_comp = 10,
  file1 = "ElbowPlot.jpeg",
  file2 = "clustering.jpeg"
)
}
\arguments{
\item{n_comp}{Dimensions of reduction to use as input}

\item{matrix}{matrix of features}
}
\value{
features_by_cells Seurat Object
}
\description{
gene_set_score_in_clusters
}
