% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions_to_genes.R
\name{regions_to_genes}
\alias{regions_to_genes}
\title{Create a map between CNV regions and genes}
\usage{
regions_to_genes(CNV = NULL, CNV_input = NULL)
}
\description{
Create a map between CNV regions and genes
}
