% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_data_bin.R
\name{sc_data_bin}
\alias{sc_data_bin}
\title{sc_data_bin}
\usage{
sc_data_bin(genes_by_cells, nbins = 25, na.rm = FALSE)
}
\arguments{
\item{genes_by_cells}{genes-by-cells matrix}

\item{nbins}{number of bins}

\item{na.rm}{whether to remove NA or not}
}
\value{
vector with bin id for each row
}
\description{
Split the input data matrix by rows into nbins
}
\details{
Split the input data matrix by rows into nbins
}
