% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scML_demo.R
\docType{data}
\name{scML_demo}
\alias{scML_demo}
\title{scML_demo
scMuffinList object with a partial dataset used in vignettes.}
\format{
An object of class \code{list} of length 9.
}
\usage{
scML_demo
}
\description{
scML_demo
scMuffinList object with a partial dataset used in vignettes.
}
\keyword{datasets}
