% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMuffin.R
\name{scMuffin}
\alias{scMuffin}
\title{scMuffin: main function that runs the whole pipeline}
\usage{
scMuffin(genes_by_cells)
}
\arguments{
\item{genes_by_cells}{genes-by-cells input matrix}
}
\description{
scMuffin: main function that runs the whole pipeline
}
