% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMuffinList_demo.R
\docType{data}
\name{scMuffinList_demo}
\alias{scMuffinList_demo}
\title{scMuffinList_demo
scMuffinList object with a partial dataset used in vignettes.}
\format{
An object of class \code{list} of length 9.
}
\usage{
scMuffinList_demo
}
\description{
scMuffinList_demo
scMuffinList object with a partial dataset used in vignettes.
}
\keyword{datasets}
