% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tissues.R
\name{show_tissues}
\alias{show_tissues}
\title{Show the tissues available for \code{\link{prepare_gsls}}}
\usage{
show_tissues(gsc_source = c("PNDB", "CM_normal", "CM_cancer"))
}
\arguments{
\item{gsc_source}{the source database}
}
\value{
List of gsc_source with tissue vectors.
}
\description{
Show the tissues available by means of \code{\link{prepare_gsls}}
}
