% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_CancerSEA.R
\docType{data}
\name{SIG_CancerSEA}
\alias{SIG_CancerSEA}
\title{CancerSEA functional states
Collected from the URL http://biocc.hrbmu.edu.cn/CancerSEA/goDownload}
\format{
An object of class \code{list} of length 14.
}
\usage{
SIG_CancerSEA
}
\description{
CancerSEA functional states
Collected from the URL http://biocc.hrbmu.edu.cn/CancerSEA/goDownload
}
\keyword{datasets}
