% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_CM_cancer.R
\docType{data}
\name{SIG_CM_cancer}
\alias{SIG_CM_cancer}
\title{signatures
SIG_CM_cancer}
\format{
An object of class \code{list} of length 278.
}
\usage{
SIG_CM_cancer
}
\description{
signatures
SIG_CM_cancer
}
\keyword{datasets}
