% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_CM_cancer_ss.R
\docType{data}
\name{SIG_CM_cancer_ss}
\alias{SIG_CM_cancer_ss}
\title{signatures
SIG_CM_cancer_ss}
\format{
An object of class \code{list} of length 15.
}
\usage{
SIG_CM_cancer_ss
}
\description{
signatures
SIG_CM_cancer_ss
}
\keyword{datasets}
