% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_CM_normal.R
\docType{data}
\name{SIG_CM_normal}
\alias{SIG_CM_normal}
\title{CellMarker Cancer markers
Collected from the URL http://bio-bigdata.hrbmu.edu.cn/CellMarker/}
\format{
An object of class \code{list} of length 1043.
}
\usage{
SIG_CM_normal
}
\description{
CellMarker Cancer markers
Collected from the URL http://bio-bigdata.hrbmu.edu.cn/CellMarker/
}
\keyword{datasets}
