% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_CSEA.R
\docType{data}
\name{SIG_CSEA}
\alias{SIG_CSEA}
\title{signatures
CancerSEA functional states}
\format{
An object of class \code{list} of length 14.
}
\usage{
SIG_CSEA
}
\description{
signatures
CancerSEA functional states
}
\keyword{datasets}
