% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_PNDB.R
\docType{data}
\name{SIG_PNDB}
\alias{SIG_PNDB}
\title{Panglao DB markers
Collected from the URL https://panglaodb.se/}
\format{
An object of class \code{list} of length 178.
}
\usage{
SIG_PNDB
}
\description{
Panglao DB markers
Collected from the URL https://panglaodb.se/
}
\keyword{datasets}
