% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_ribosomal_proteins.R
\docType{data}
\name{SIG_ribosomal_proteins}
\alias{SIG_ribosomal_proteins}
\title{Gene set of ribosomal proteins}
\format{
An object of class \code{character} of length 80.
}
\usage{
SIG_ribosomal_proteins
}
\description{
Gene set of ribosomal proteins
}
\keyword{datasets}
