% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG_Tirosh.R
\docType{data}
\name{SIG_Tirosh}
\alias{SIG_Tirosh}
\title{Cell cycle gene sets
Collected from Tirosh, et al., Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science. 2016. doi: 10.1126/science.aad0501.}
\format{
An object of class \code{list} of length 2.
}
\usage{
SIG_Tirosh
}
\description{
Cell cycle gene sets
Collected from Tirosh, et al., Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science. 2016. doi: 10.1126/science.aad0501.
}
\keyword{datasets}
