% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{signatures
CancerSEA functional states}
\format{
An object of class \code{list} of length 14.
}
\usage{
signatures
}
\description{
signatures
CancerSEA functional states
}
\keyword{datasets}
