% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clusterings.R
\name{add_clusterings}
\alias{add_clusterings}
\title{Adds a clustering to the clustering list}
\usage{
add_clusterings(clusterings = NULL, cell_id = NULL, values = NULL)
}
\arguments{
\item{clusterings}{clustering list}

\item{cell_id}{vector of cell identifiers}

\item{values}{vector or data.frame with clustering labels}
}
\description{
Adds a clustering to the clustering list
}
