% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_features}
\alias{add_features}
\title{Add a feature to scMuffinList}
\usage{
add_features(scMuffinList = NULL, name = NULL, summary = NULL, full = NULL)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{name}{feature name}

\item{summary}{data.frame with cell identifiers as row.names}

\item{full}{optional information related to the feature, like cell-level statistics}
}
\value{
scMuffinList with the additional feature
}
\description{
Add a feature to scMuffinList
}
