% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_partitions.R
\name{add_partitions}
\alias{add_partitions}
\title{Adds a clustering to the clustering list}
\usage{
add_partitions(partitions_obj = NULL, to_add = NULL)
}
\arguments{
\item{partitions_obj}{clusterings object}

\item{to_add}{data.frame of 1 or more cell labels with cell id as row names}
}
\description{
Adds a clustering to the clustering list
}
