% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_partitions.R
\name{add_partitions}
\alias{add_partitions}
\title{Adds a clustering to the clustering list}
\usage{
add_partitions(scMuffinList = NULL, clusters = NULL, partition_id = NULL)
}
\arguments{
\item{scMuffinList}{clusterings object}

\item{clusters}{named factor with cluster label for each cell}

\item{partition_id}{partition id}
}
\description{
Adds a clustering to the parition list
}
