% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_outliers_col.R
\name{adj_outliers_col}
\alias{adj_outliers_col}
\title{Remove outliers to avoid that these influence colors}
\usage{
adj_outliers_col(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Remove outliers to avoid that these influence colors
}
