% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_cluster_enrichment.R
\name{assess_cluster_enrichment}
\alias{assess_cluster_enrichment}
\title{assess_cluster_enrichment}
\usage{
assess_cluster_enrichment(
  features,
  partitions,
  meta_clusters = FALSE,
  write_output = TRUE,
  out_dir = "./"
)
}
\arguments{
\item{features}{feature list}

\item{partitions}{clustering list}

\item{meta_clusters}{boolean, TRUE if the assessment involves meta clusters}

\item{write_output}{boolean, TRUE to write the output in the output dir}

\item{out_dir}{output dir}
}
\description{
assess_cluster_enrichment
}
