% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_cluster_enrichment.R
\name{assess_cluster_enrichment}
\alias{assess_cluster_enrichment}
\title{Assess enrichment with ORA or GSEA}
\usage{
assess_cluster_enrichment(
  scMuffinList = NULL,
  feature_name = NULL,
  partition_id = NULL,
  meta_clusters = FALSE
)
}
\arguments{
\item{meta_clusters}{boolean, TRUE if the assessment involves meta clusters}

\item{features}{feature list}

\item{partitions}{clustering list}
}
\description{
Assess cluster enrichment
}
