% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_cluster.R
\name{barplot_cluster}
\alias{barplot_cluster}
\title{Barplot of cluster enrichment in a categorical feature}
\usage{
barplot_cluster(
  scMuffinList = NULL,
  feature_name = NULL,
  feature_id = NULL,
  partition_id = NULL,
  dir_out = "./",
  only_pos_nes = TRUE,
  cex.axis = 0.8,
  p.type = c("p", "p_adj"),
  width = 180,
  height = 180,
  units = "mm",
  res = 300,
  legend.pos = "right"
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{feature_name}{name of the features to be used.}

\item{feature_id}{identifier of the feature to be used.}

\item{partition_id}{one among the partitions}

\item{dir_out}{string, output directory}

\item{only_pos_nes}{whether to consider only positive enrichments}

\item{cex.axis}{cex.axis}

\item{p.type}{p for nominal p-value or p_adj for BH FDR.}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{legend.pos}{position of barplot legend}
}
\description{
Produce barplots (1 for each cluster) of distribution of cells associated with the values of the selected feature. A png figure for each cluster is saved in dir_out.
}
