% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster.R
\name{boxplot_cluster}
\alias{boxplot_cluster}
\title{Bocplt of cluster enrichment in a quantitative feature}
\usage{
boxplot_cluster(
  scMuffinList = NULL,
  feature_name = NULL,
  partition_id = NULL,
  dir_out = "./",
  n_features = 10,
  only_pos_nes = TRUE,
  do_scale_features = FALSE,
  cex.axis = 0.8,
  width = 180,
  height = 180,
  units = "mm",
  res = 300
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{feature_name}{the names of the feature that should be considered. It must be one of names(scMuffinList)}

\item{partition_id}{one among the partitions}

\item{dir_out}{output directory}

\item{n_features}{maximum number of features that will be shown}

\item{only_pos_nes}{whether to consider only positive enrichments}

\item{do_scale_features}{whether to scale features}

\item{cex.axis}{cex.axis}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}
}
\description{
Produce boxplots to visualize the distribution of cell values according to the selected figure. A png figure for each cluster is saved in dir_out.
}
