% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster.R
\name{boxplot_cluster}
\alias{boxplot_cluster}
\title{boxplot_cluster}
\usage{
boxplot_cluster(
  features_by_cells_matrix,
  cell_clusters,
  n_top = 3,
  dir_out = "./"
)
}
\description{
boxplot_cluster
}
