% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster.R
\name{boxplot_cluster}
\alias{boxplot_cluster}
\title{Boxplot clusters 
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.}
\usage{
boxplot_cluster(
  features = NULL,
  clusterings = NULL,
  clustering_name = NULL,
  clust_enrich_res = NULL,
  dir_out = "./",
  only_top = 10,
  criterion = "fdr",
  fdr_threshold = NULL,
  only_pos_nes = TRUE,
  do_scale_features = FALSE,
  ...
)
}
\arguments{
\item{features}{feature list}

\item{clusterings}{clusterings object}

\item{clustering_name}{one among the clusterings}

\item{clust_enrich_res}{clustering enrichment results}

\item{dir_out}{output directory}

\item{only_top}{numeric, number of features to be shown with a different color and representing the most significative features according to t-test}

\item{criterion}{"fdr" to sort features by fdr}

\item{fdr_threshold}{fdr threshold}

\item{only_pos_nes}{whether to consider only positive enrichments}

\item{do_scale_features}{whether to scale features}

\item{...}{passed to axis}
}
\description{
Boxplot clusters 
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
