% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_cluster.R
\name{boxplot_cluster}
\alias{boxplot_cluster}
\title{Boxplot clusters}
\usage{
boxplot_cluster(
  features = NULL,
  cell_clusters = NULL,
  cluster_enrichment = NULL,
  dir_out = "./",
  only_top = 10,
  criterion = "fdr",
  fdr_threshold = NULL,
  only_pos_nes = TRUE,
  ...
)
}
\arguments{
\item{cell_clusters}{array with cell values named by their cluster ID}

\item{dir_out}{string, output directory}

\item{...}{passed to axis()}

\item{cells_by_features}{matrix, features by cells matrix}

\item{n_top}{numeric, number of features to be shown with a difference color and representing the most significative features according to t-test}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Noemi Di Nanni
}
