% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gs_perm.R
\name{calc_gs_perm}
\alias{calc_gs_perm}
\title{Calculate permutations}
\usage{
calc_gs_perm(rll, perm, gs)
}
\arguments{
\item{rll}{numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory}

\item{perm}{vector of permuted names}

\item{gs}{gene set}
}
\description{
Calculate permutations
}
