% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CNV.R
\name{calculate_CNV}
\alias{calculate_CNV}
\title{Calculate CNV}
\usage{
calculate_CNV(genes_by_cells, reference = NULL, mc.cores = 2)
}
\arguments{
\item{reference}{NULL. It is possible to add a reference vector, downloaded from GTEx portal.}

\item{genes_by_cell}{Preprocessed Seurat object}
}
\description{
Calculate CNV
}
