% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CNV.R
\name{calculate_CNV}
\alias{calculate_CNV}
\title{Calculate CNV}
\usage{
calculate_CNV(
  genes_by_cells,
  reference = NULL,
  mc.cores = 1,
  wnd_size = 100,
  min_genes = 1000,
  min_cells = 100,
  expr_lim = NULL,
  scale_cells = TRUE,
  na.rm = FALSE,
  center_genes = FALSE
)
}
\arguments{
\item{genes_by_cells}{genes-by-cells data.frame}

\item{reference}{one-column data.frame with a reference expression profile; rownames must match those of genes_by_cell;}

\item{mc.cores}{number of cores;}

\item{wnd_size}{number of adjacent genes considered;}

\item{min_genes}{minimun number of genes expressed in a cell;}

\item{min_cells}{minimum numbver of cells in which a gene must be expressed;}

\item{expr_lim}{min and max values of relative expression; by default, lower and higher whiskers returned by grDevices::boxplot.stats will be used. Set to NA or anything other value such that length(expr_lim) != 2 to disbale the removal of outliers.}

\item{scale_cells}{whether to scale cells}

\item{na.rm}{whether to remove 0 values in CNV estimation}

\item{center_genes}{whether to center genes or not}
}
\description{
Calculate CNV using the moving average approach firstly described in Patel et al., 2014 Science (DOI: 10.1126/science)
}
