% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gs_scores_in_clusters.R
\name{calculate_gs_scores_in_clusters}
\alias{calculate_gs_scores_in_clusters}
\title{calculate_signatures_clusters}
\usage{
calculate_gs_scores_in_clusters(
  gs_scores_obj = NULL,
  cell_clusters = NULL,
  ncells_min = 5,
  null_model = TRUE
)
}
\arguments{
\item{gs_scores_obj}{list of gene signature lists}

\item{cell_clusters}{cell cluster labels}

\item{ncells_min}{min number of cells required for the clculation of the average signature in the cluster}

\item{null_model}{TRUE to consider the permutations}
}
\description{
calculate_signatures_clusters
}
