% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gs_scores_in_clusters.R
\name{calculate_gs_scores_in_clusters}
\alias{calculate_gs_scores_in_clusters}
\title{Calculate gene set score at cluster level}
\usage{
calculate_gs_scores_in_clusters(
  scMuffinList = NULL,
  partition_id = NULL,
  ncells_min = 5,
  null_model = TRUE,
  alt = "g",
  test = "t"
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{partition_id}{identifier of the partition to be used}

\item{ncells_min}{minimum number of cells required for the calculation of the average signature in the cluster}

\item{null_model}{TRUE to consider the empirical null based on gene set permutations}

\item{alt}{alterative passed to [wilcox.test()] or [t.test()]}

\item{test}{type of test: t to use [t.test()]; wrs to use [wilcox.test()]}
}
\value{
scMuffinList with cluster level scores in `sMuffinList$cluster_data[[partition_id]]`. The element [summary] contains a clusters-by-gene sets table, while the element [full] the full result
}
\description{
Calculate gene set score at cluster level
}
