% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signatures.R
\name{calculate_signatures}
\alias{calculate_signatures}
\title{calculate signatures}
\usage{
calculate_signatures(
  genes_by_cells,
  custom_signatures = NULL,
  mc.cores = 2,
  cell_clusters = NULL,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50
)
}
\description{
calculate signatures
}
