% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signatures_clusters.R
\name{calculate_signatures_clusters}
\alias{calculate_signatures_clusters}
\title{calculate_signatures_clusters}
\usage{
calculate_signatures_clusters(
  sign_list,
  cell_clusters = cell_clusters,
  ncells_min = 5,
  null_model = TRUE
)
}
\description{
calculate_signatures_clusters
}
