% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_by_features.R
\name{cluster_by_features}
\alias{cluster_by_features}
\title{Cluster by features}
\usage{
cluster_by_features(
  features,
  n_comp = 10,
  cnv = FALSE,
  plot_umap = FALSE,
  out_dir = "./",
  ...
)
}
\arguments{
\item{n_comp}{numeric, Dimensions of reduction to use as input}

\item{cnv}{TRUE/FALSE set it to TRUE for clustering CNV results}

\item{features_by_cells}{matrix, features by cells matrix}
}
\value{
features_by_cells Seurat Object, object with saved dimension reduction components calculate on features by cells matrix
}
\description{
Clustering of the features_by_cell matrix
}
