% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_by_gs.R
\name{cluster_by_gs}
\alias{cluster_by_gs}
\title{cluster by n-genes}
\usage{
cluster_by_gs(Seu_obj, gs = NULL)
}
\arguments{
\item{Seu_obj}{Seu_obj}

\item{gs}{vector of genes}
}
\description{
cluster by n-genes
}
