% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gsea.R
\name{cluster_gsea}
\alias{cluster_gsea}
\title{Calculate cluster enrichment by gsea approach}
\usage{
cluster_gsea(feature_values, cell_clusters, min.cells = 100, mc.cores = 1)
}
\arguments{
\item{feature_values}{matrix, features by cells matrix}

\item{cell_clusters}{array with cell values named by their cluster ID}

\item{min.cells}{minimum number of cells}

\item{mc.cores}{number of cores}
}
\description{
Calculate cluster enrichment by gsea approach
}
\details{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Ettore Mosca
}
