% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_hyper.R
\name{cluster_hyper}
\alias{cluster_hyper}
\title{Calculate cluster enrichment by hypergeometric test}
\usage{
cluster_hyper(feature_values, cell_clusters, fdr = 0.05, top = 2)
}
\arguments{
\item{feature_values}{matrix, features by cells matrix}

\item{cell_clusters}{array with cell values named by their cluster ID}

\item{fdr}{fdr threshold, default at 0.05}

\item{top}{numeric, number of features to be shown with a difference color and representing the most significative features according to t-test}

\item{mc.cores}{number of cores}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\details{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
\author{
Ettore Mosca
}
