% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Cell cluster statistics}
\usage{
cluster_stats(
  scMuffinList = NULL,
  partition_id = NULL,
  feature_id = NULL,
  mean_f = mean,
  var_f = sd,
  na.rm = TRUE
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{partition_id}{partition id}

\item{feature_id}{feature id}

\item{na.rm}{whether to remove na or not}

\item{mean}{function to use for average}

\item{var}{function to use for variability}
}
\value{
scMuffinList with data.frames "mean" and "var" added to "scMuffinList$cluster_data[[partition_id]]"
}
\description{
Calculate mean and variance for partition id and feature id
}
