% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_stats}
\alias{cluster_stats}
\title{cluster_stats}
\usage{
cluster_stats(
  feat_obj = NULL,
  clusterings = NULL,
  which_col = NULL,
  avg = mean,
  var = sd,
  na.rm = TRUE
)
}
\arguments{
\item{feat_obj}{feature object}

\item{clusterings}{clusterings object}

\item{which_col}{which column of the clustering object should be used. if NULL, all columns will be considered}

\item{avg}{function to use for average}

\item{var}{function to use for variability}

\item{na.rm}{whether to remove na or not}
}
\description{
cluster_stats
}
