% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV.R
\name{CNV}
\alias{CNV}
\title{CNV}
\usage{
CNV(x, wnd_size = 100, na.rm = FALSE)
}
\arguments{
\item{x}{list of dataframes retrieved by 'preprocess_object_for_cnv'.}
}
\description{
Function to be used in calculate_CNV.
}
\references{
"Single-cell RNA-seq highlights intratumoral heterogeneity in primary glioblastoma" Patel et al., Supplementary materials
}
\author{
Valentina Nale
}
