% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_analysis.R
\name{CNV_analysis}
\alias{CNV_analysis}
\title{CNV analysis}
\usage{
CNV_analysis(
  scMuffinList = NULL,
  mc.cores = 1,
  reference = NULL,
  min_cells = 100,
  min_genes = 100,
  wnd_size = 100,
  scale_cells = TRUE,
  center_genes = FALSE,
  expr_lim = FALSE,
  method = "mean",
  na.rm = FALSE,
  z.score = FALSE,
  eps = NULL
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{mc.cores}{number of cores;}

\item{reference}{one-column data.frame with a reference expression profile; rownames must match those of genes_by_cell;}

\item{min_cells}{minimum numbver of cells in which a gene must be expressed;}

\item{min_genes}{minimun number of genes expressed in a cell;}

\item{wnd_size}{number of adjacent genes considered;}

\item{scale_cells}{whether to scale cells}

\item{center_genes}{whether to center genes or not}

\item{expr_lim}{min and max values of relative expression; by default, lower and higher whiskers returned by grDevices::boxplot.stats will be used. Set to NA or anything other value such that length(expr_lim) != 2 to disable the removal of outliers.}

\item{method}{mean: subtract the average profile of the reference cluster to every cell; min_max see Tirosh et al.}

\item{na.rm}{whether to remove 0 values in CNV estimation}

\item{z.score}{whether to use z-scores of the cluster median CNV profile, instead of the median itself.}

\item{eps}{absolute threshold to call CNV regions.}
}
\description{
CNV analysis
}
