% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_clusterings.R
\name{create_clusterings}
\alias{create_clusterings}
\title{Create the clustering object}
\usage{
create_clusterings(x)
}
\arguments{
\item{x}{data.frame of 1 or more cell labels with cell id as row names}
}
\description{
Create the clustering object
}
