% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_features.R
\name{create_features}
\alias{create_features}
\title{Create feature list}
\usage{
create_features(cell_id = NULL, values = NULL)
}
\arguments{
\item{cell_id}{vector of cell id}

\item{values}{vector of values}
}
\description{
Create feature list
}
