% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scMuffinList.R
\name{create_scMuffinList}
\alias{create_scMuffinList}
\title{Create scMuffinList}
\usage{
create_scMuffinList(counts = NULL, normalized = NULL)
}
\arguments{
\item{counts}{genes-by-cell matrix or data.frame of counts.}

\item{normalized}{genes-by-cell matrix or data.frame with normalized expression.}
}
\value{
scMuffinList with counts and normalized elements as dgCMatrix.
}
\description{
Create scMuffinList from counts and/or normalized data.
}
