% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csea.R
\name{csea}
\alias{csea}
\title{Gene Set Enrichment Analysis}
\usage{
csea(
  rl,
  gsl,
  k = 100,
  min.size = 100,
  ord.mode = -1,
  min.k.nes = 10,
  mc_cores_path = 1,
  mc_cores_perm = 1
)
}
\arguments{
\item{rl}{numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory}

\item{gsl}{named list of gene sets}

\item{k}{number of permutations}

\item{ord.mode}{ordering mode: -1 -> descending; 1 ascending; must be of length equal to `ncol(rl)`}

\item{mc_cores_path}{number of cores to use for parallel calculation of gene set lists; the total number of cpu used will be mc_cores_path x mc_cores_perm}

\item{mc_cores_perm}{number of cores to use for parallel calculation of ranked list permutations; the total number of cpu used will be mc_cores_path x mc_cores_perm}
}
\value{
list with two data.frames, gs_table and leading_edge. gs_table contains: es, enrichment score; nes normalized enrichment score; p-value, empirical p-value; adjusted p-value, BH FDR; FDR q-value, empirical FDR. leading_edge contains: tags, leading edge size; tags_perc, leading edge size percent over gene set; list_top, rank of the ES; list_top_perc, rank of the ES percent over full ranked list; lead_edge, gene names of the leading edge.
}
\description{
Gene Set Enrichment Analysis
}
