% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csea.R
\name{csea}
\alias{csea}
\title{Gene Set Enrichment Analysis}
\usage{
csea(
  rl,
  gsl,
  k = 100,
  min.size = 100,
  ord.mode = -1,
  min.k.nes = 10,
  mc_cores_path = 1,
  mc_cores_perm = 1
)
}
\arguments{
\item{rl}{numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory}

\item{gsl}{named list of gene sets}

\item{k}{number of permutations}

\item{min.size}{minimum number of cells with a not null value}

\item{ord.mode}{ordering mode: -1 -> descending; 1 ascending; must be of length equal to `ncol(rl)`}

\item{min.k.nes}{minimum number of not null NES values}

\item{mc_cores_path}{number of cores to use for parallel calculation of gene set lists; the total number of cpu used will be mc_cores_path * mc_cores_perm}

\item{mc_cores_perm}{number of cores to use for parallel calculation of ranked list permutations; the total number of cpu used will be mc_cores_path * mc_cores_perm}
}
\value{
list with two data.frames, gs_table and leading_edge.
}
\description{
Gene Set Enrichment Analysis
}
