% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_CNV_regions.R
\name{detect_CNV_regions}
\alias{detect_CNV_regions}
\title{Detect CNV regions}
\usage{
detect_CNV_regions(scMuffinList = NULL, z.score = FALSE, eps = NULL)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{z.score}{whether to use clustere median z-scores of CNV signal (TRUE) or cluster median CNV signal (FALSE).}

\item{eps}{the value that defines a CNV. If NUll the stard deviation of the dataset is used.}
}
\description{
Plot an heatmap of the CNV.
}
\details{
CNV Profile of every cluster
}
