% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_map.R
\name{diff_map}
\alias{diff_map}
\title{Computation of Diffusion Map}
\usage{
diff_map(genes_by_cells = NULL, output_ndc = 1:3, root_cell = NULL, ...)
}
\arguments{
\item{genes_by_cells}{normalized genes-by-cells expression matrix}

\item{output_ndc}{The numnber of diffusion components that will be included in the output}

\item{root_cell}{root cell or "random" to pick up a random root via the destiny::random_root function}

\item{...}{parameters for destiny::DiffusionMap}
}
\description{
The function apply Landscent package to a gene by cell matric to compute signal entropy, potency states and diffusion pseudo time
}
