% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_map.R
\name{diff_map}
\alias{diff_map}
\title{Computation of Diffusion Map and Diffusion Pseudo Time}
\usage{
diff_map(scMuffinList = NULL, root_cell = "random", n_pcs = 50, ...)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{root_cell}{root cell or "random" to pick up a random root via the destiny::random_root function}

\item{n_pcs}{number of PCs to consider, 50 by default}

\item{...}{parameters passed to destiny::DiffusionMap}
}
\value{
scMuffinList with element "diffusion_map_pseudo_t", a list with summary and full. summary is a data.frame with: the first two eigenvectors; branch, branch label for each cell; tips, whether the cell is a tip of the branching. full contains dpt object, generated by [destiny::DPT()]. See destiny for further information.
}
\description{
The functions computes a Diffusion Map and Diffusion Pseudo Time using package destiny.
}
