% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_cluster.R
\name{dotplot_cluster}
\alias{dotplot_cluster}
\title{Boxplot clusters}
\usage{
dotplot_cluster(
  features,
  partitions = NULL,
  clustering_name = NULL,
  dir_out = "./",
  clust_enrich_res = NULL,
  text_val = c("p", "wbd", "er")
)
}
\arguments{
\item{features}{feautures object}

\item{partitions}{partitions object}

\item{clustering_name}{one among the partitions}

\item{dir_out}{string, output directory}

\item{clust_enrich_res}{clustering enrichment results}

\item{text_val}{which column of the hypergeometric result. Default is "p". Other values are "wbd" and "er"}
}
\description{
Produce boxplots of the given features in each cluster. A t-test is performed for each feature among clusters.
}
