% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_Rate.R
\name{exp_Rate}
\alias{exp_Rate}
\title{Expression_rate}
\usage{
exp_Rate(genes_by_cells, ngenes_min = 5)
}
\arguments{
\item{genes_by_cells}{real data}

\item{ngenes_min}{min number of genes in each cell}
}
\description{
Expression_rate
}
