% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_rate.R
\name{exp_rate}
\alias{exp_rate}
\title{Expression_rate
ration between the number of expressed genes and the total number of reads}
\usage{
exp_rate(genes_by_cells, ngenes_min = 5)
}
\arguments{
\item{genes_by_cells}{real data}

\item{ngenes_min}{min number of genes in each cell}
}
\value{
expRate_vect vector with expression rate for each cell
}
\description{
Expression_rate
ration between the number of expressed genes and the total number of reads
}
