% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_table.R
\name{extract_cluster_enrichment_table}
\alias{extract_cluster_enrichment_table}
\title{Extract cluster enrichment table}
\usage{
extract_cluster_enrichment_table(
  scMuffinList,
  partition_id = NULL,
  type = NULL,
  quantity = NULL,
  feature_id = NULL
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{partition_id}{id of the partition to be used}

\item{type}{"CSEA" or "ORA"}

\item{quantity}{possible values are as follows. CSEA: es, enrichment score; p_val empirical p-value; adj_p_val, FDR (BH); nes normalized enrichment score; FDRq empirical FDR. ORA: wbd (white balls drawn), number of cells with the value of interest in the cluster; exp expected number of cells with the value of interest in the cluster; er, enrichment ratio; p, hypergeometric test p-value; p_adj, BH FDR}

\item{feature_id}{name of the features that will be considered. Must be any of `colnames(scMuffinList$cluster_data$[[parition_id]]$CSEA)` or `names(scMuffinList$cluster_data$[[parition_id]]$ORA)`}
}
\value{
A cluster enrichment table with the desired quantity
}
\description{
Extract cluster enrichment table with the desired quantity from the results generated by means of [assess_cluster_enrichment()]
}
