% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_table.R
\name{extract_cluster_enrichment_table}
\alias{extract_cluster_enrichment_table}
\title{extract_cluster_enrichment_table}
\usage{
extract_cluster_enrichment_table(
  clust_enrich_res,
  q_type = "nes",
  c_type = "er"
)
}
\arguments{
\item{clust_enrich_res}{cluster enrichment result}

\item{q_type}{the column of GSEA result that will appear in the output table}

\item{c_type}{column of hyper result that will appear in the output table}
}
\description{
Extract cluster enrichment table
}
