% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_table.R
\name{extract_cluster_enrichment_table}
\alias{extract_cluster_enrichment_table}
\title{extract_cluster_enrichment_table}
\usage{
extract_cluster_enrichment_table(
  clust_enrich_res,
  q_type = "nes",
  c_type = "er"
)
}
\arguments{
\item{features}{feature list}

\item{clusterings}{clustering list}

\item{meta_clusters}{boolean, TRUE if the assessment involves meta clusters}

\item{write_output}{boolean, TRUE to write the output in the output dir}

\item{out_dir}{output dir}
}
\description{
extract_cluster_enrichment_table
}
