% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_tags.R
\name{extract_cluster_enrichment_tags}
\alias{extract_cluster_enrichment_tags}
\title{extract_cluster_enrichment_table}
\usage{
extract_cluster_enrichment_tags(
  clust_enrich_res,
  q_selection_criterion = "FDRq",
  q_selection_threshold = 0.1,
  q_sort_crit = "nes",
  q_sort_desc = TRUE,
  only_pos_nes = TRUE,
  c_selection_criterion = "p_adj",
  c_selection_threshold = 0.1,
  c_sort_crit = "p",
  c_sort_desc = FALSE
)
}
\arguments{
\item{clust_enrich_res}{cluster enrichment result}

\item{q_selection_criterion}{criteria for selection of tags from CSEA}

\item{q_selection_threshold}{threshold for selection of tags from CSEA}

\item{q_sort_crit}{criteria for sorting tags from CSEA}

\item{q_sort_desc}{sorting in decreasing order or not}

\item{only_pos_nes}{only positive nes?}

\item{c_selection_criterion}{criteria for selection of tags from ORA}

\item{c_selection_threshold}{threshold for selection of tags from ORA}

\item{c_sort_crit}{criteria for sorting tags from ORA}

\item{c_sort_desc}{sorting in decreasing order or ORA}
}
\description{
Extract cluster enrichment table
}
