% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster_enrichment_tags.R
\name{extract_cluster_enrichment_tags}
\alias{extract_cluster_enrichment_tags}
\title{Extract the top results of GSEA and ORA for each cluster}
\usage{
extract_cluster_enrichment_tags(
  scMuffinList = NULL,
  partition_id = NULL,
  GSEA_selection_criterion = "FDRq",
  GSEA_selection_threshold = 0.25,
  only_pos_nes = TRUE,
  ORA_selection_criterion = "p_adj",
  ORA_selection_threshold = 0.1,
  n_max_per_cluster = 3
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{partition_id}{one among the partitions}

\item{GSEA_selection_criterion}{selection criteria for results of GSEA}

\item{GSEA_selection_threshold}{threshold for selection of tags from GSEA}

\item{ORA_selection_criterion}{selection criteria for results of ORA}

\item{ORA_selection_threshold}{threshold for selection of tags from ORA}

\item{n_max_per_cluster}{maximum number of tags per cluster}
}
\value{
Add or overwrite the "cluster_tags" list to `scMuffinList$cluster_data[[partition_id]]`
}
\description{
Extract the best results of GSEA and ORA for each cluster
}
