% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_score.R
\name{gene_set_score}
\alias{gene_set_score}
\title{gene_set_score}
\usage{
gene_set_score(
  gene_set_data,
  control_set_data,
  nmark_min = 5,
  ncells_min = NULL
)
}
\arguments{
\item{gene_set_data}{real data}

\item{control_set_data}{control data}

\item{nmark_min}{min number of markers}

\item{ncells_min}{min number of cells}
}
\description{
gene_set_score
}
