% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_score.R
\name{gene_set_score}
\alias{gene_set_score}
\title{gene_set_score}
\usage{
gene_set_score(
  gene_set,
  genes_by_cells,
  bins,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50,
  verbose = TRUE,
  null_model = TRUE
)
}
\arguments{
\item{genes_by_cells}{real data}

\item{bins}{random data}

\item{nmark_min}{min number of markers}

\item{ncells_min}{min number of cells}

\item{k}{number of random gene set to be considered}
}
\value{
list with
score_table: data frame with gene set score for each cell
permutations: list with the permutations
}
\description{
gene_set_score
}
\author{
Ettore Mosca
}
