% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_score_in_clusters.R
\name{gene_set_score_in_clusters}
\alias{gene_set_score_in_clusters}
\title{gene_set_score_in_clusters}
\usage{
gene_set_score_in_clusters(
  score_table,
  cell_clusters,
  ncells_min = 5,
  alt = "g",
  test = "t",
  null_model = TRUE
)
}
\arguments{
\item{score_table}{real data}

\item{cell_clusters}{vector of cell clusters}

\item{ncells_min}{minimum number of cells}

\item{alt}{alterative}

\item{test}{type of test}

\item{null_model}{whether to use permutations or not}
}
\description{
gene_set_score_in_clusters
}
