% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_score_in_clusters.R
\name{gene_set_score_in_clusters}
\alias{gene_set_score_in_clusters}
\title{gene_set_score_in_clusters}
\usage{
gene_set_score_in_clusters(
  score_table,
  cell_clusters,
  ncells_min = 5,
  alt = "g",
  test = "t"
)
}
\arguments{
\item{alt}{alterative}

\item{test}{type of test}

\item{current_gene_set}{real data}

\item{seurat_object_data}{control data}

\item{seurat_object_ident}{min number of markers}
}
\description{
gene_set_score_in_clusters
}
