% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_score.R
\name{gs_score}
\alias{gs_score}
\title{gene_set_score}
\usage{
gs_score(
  gene_set = NULL,
  genes_by_cells = NULL,
  bins = NULL,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50,
  verbose = TRUE,
  null_model = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{gene_set}{gene set}

\item{genes_by_cells}{real data}

\item{bins}{random data}

\item{nmark_min}{min number of markers}

\item{ncells_min}{min number of cells}

\item{k}{number of random gene set to be considered}

\item{kmin}{minimim number of k}

\item{verbose}{verbosity}

\item{null_model}{boolean, whether to use the permutations or not}
}
\value{
list with
   score_table: data frame with gene set score for each cell
   permutations: list with the permutations
}
\description{
Gene set scoring
}
\author{
Ettore Mosca
}
