% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_score.R
\name{gs_score}
\alias{gs_score}
\title{Gene set scoring}
\usage{
gs_score(
  gene_set = NULL,
  genes_by_cells = NULL,
  bins = NULL,
  nmark_min = 5,
  ncells_min = 5,
  k = 100,
  kmin = 50,
  verbose = FALSE,
  null_model = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{gene_set}{gene set}

\item{genes_by_cells}{real data}

\item{bins}{random data}

\item{nmark_min}{min number of markers}

\item{ncells_min}{min number of cells}

\item{k}{number of random gene set to be considered}

\item{kmin}{minimim number of k}

\item{verbose}{verbosity}

\item{null_model}{boolean, whether to use the permutations or not}

\item{na.rm}{whether to remove or not the missing values}
}
\value{
A data.frame with the following columns: case, the gene set average observed in the cell; case.N, number of genes in the gene set; case.AV, number of genes available in the cell; nmark_min, whether case.AV > nmark_min; avg_control, average of the averag gene set score in control cells; control.AV, number of available control cells, null_ok, whether a null could be defined for the cell; avg_delta_score, average of case minu each control cell; delta_score, case - avg_control.
}
\description{
Calculate gene set scores using the approach described in Tirosh2016
}
\references{
Tirosh2016 10.1126/science.aad0501
}
