% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_scores_in_clusters.R
\name{gs_scores_in_clusters}
\alias{gs_scores_in_clusters}
\title{gene_set_score_in_clusters}
\usage{
gs_scores_in_clusters(
  score_table = NULL,
  cell_clusters = NULL,
  ncells_min = 5,
  alt = "g",
  test = "t",
  null_model = TRUE
)
}
\arguments{
\item{score_table}{real data}

\item{cell_clusters}{vector of cell clusters}

\item{ncells_min}{minimum number of cells required for the calculation of the average signature in the cluster}

\item{alt}{alterative passed to [wilcox.test()] or [t.test()]}

\item{test}{type of test: t to use [t.test()]; wrs to use [wilcox.test()]}

\item{null_model}{TRUE to consider the empirical null based on gene set permutations}
}
\description{
Gene set scoring in clusters
}
