% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_CNV.R
\name{heatmap_CNV}
\alias{heatmap_CNV}
\title{heatmap_CNV}
\usage{
heatmap_CNV(
  cnv,
  cnv_clustering,
  ref_cluster = NULL,
  file = "heatmap_CNV.jpg",
  ...
)
}
\arguments{
\item{cnv}{CNV matrix}

\item{cnv_clustering}{cell clustering by CNV}

\item{ref_cluster}{reference cluster}

\item{file}{output file}

\item{...}{arguments passed to ComplexHeatmap::Heatmap}
}
\description{
Plot an heatmap of the CNV.
}
\details{
Preprocessing with 'preprocess_for_heatmap' needed.
}
\author{
Valentina Nale
}
