% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_for_heatmap.R
\name{heatmap_cnv}
\alias{heatmap_cnv}
\title{preprocess_for_heatmap}
\usage{
preprocess_for_heatmap(result_cnv)
}
\arguments{
\item{result_cnv}{}
}
\value{
A matrix to be used by the function heatmap_CNV.
}
\description{
Bind chromosomes together as a matrix.
}
\details{
Every chromosome is first transposed.
}
\author{
Valentina Nale
}
