% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_CNV.R
\name{heatmap_CNV}
\alias{heatmap_CNV}
\title{heatmap_CNV}
\usage{
heatmap_CNV(chr_merged)
}
\arguments{
\item{chr_merged}{...}

\item{list_ncol_chr}{...}
}
\description{
Plot as an heatmap the resulting data from CNV.
}
\details{
Preprocessing with 'preprocess_for_heatmap' needed.
}
\author{
Valentina Nale
}
