% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_datasets_comparison.R
\name{inter_dataset_comparison}
\alias{inter_dataset_comparison}
\title{Inter-dataset cluster similarity}
\usage{
inter_dataset_comparison(
  seu_obj_list = NULL,
  gsl = NULL,
  genes_min = 3,
  genes_max = 500,
  mc.cores = 1,
  null_model = TRUE,
  ncells_min = 5,
  do_plot = FALSE,
  outfile = NULL,
  pal = NULL,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  top_genes = FALSE,
  ...
)
}
\arguments{
\item{genes_min}{minimum number of genes required among the markers of a cluster}

\item{genes_max}{maximum number of genes required among the markers of a cluster}

\item{mc.cores}{number of cores}

\item{null_model}{whether to use or not the empirical null model. See calculate_signature}

\item{ncells_min}{minim number of cells in a cluster}

\item{do_plot}{whether to plot or not the results}

\item{outfile}{out file name}

\item{pal}{palette for the output heatmap}

\item{cluster_rows}{whether to cluster or not the rows}

\item{cluster_columns}{whether to cluster or not the columns}

\item{top_genes}{If specified, only the first top_genes genes of every element of the lists cluster_markers_1 and cluster_markers_2 will be considered. This implies that the markers for each clusters are considered sorted by decreasing relevance. Default is FALSE, which}

\item{...}{arguments passed to calculate_signatures}

\item{gbc_1}{genes-by-cells expression matrix of dataset 1}

\item{gbc_2}{genes-by-cells expression matrix of dataset 2}

\item{clusters_1}{named vector with cell clustering fo dataset 1}

\item{clusters_2}{named vector with cell clustering fo dataset 2}

\item{cluster_markers_1}{list of markers for each cluster of dataset 1}

\item{cluster_markers_2}{list of markers for each cluster of dataset 2}

\item{dataset_name_1}{dataset 1 name}

\item{dataset_name_2}{dataset 2 name}
}
\value{
A list with:
\itemize{
  \item{clust_sim, matrix of clucster similarity;}
  \item{m1=m1}
  \item{m2=m2}
  \item{markers_1, markers of dataset 1;}
  \item{markers_2, markers of dataset 2;}
}
}
\description{
Quantify the similarity between clusters of two datasets, on the basis of the average cluster marker expression
}
