% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscent.R
\name{landscent}
\alias{landscent}
\title{Landscent}
\usage{
landscent(
  genes_by_cells_matrix,
  ppi = NULL,
  reduceMethod = "PCA",
  clusterMethod = "dbscan",
  mc.cores = 2
)
}
\arguments{
\item{genes_by_cells_matrix}{real data}

\item{ppi}{The adjacency matrix of a user-given PPI network with rownames and colnames labeling genes (same gene identifier as in genes_by_cells_matrix)}

\item{reduceMethod}{indicates the method to do dimension reduction: "PCA" or "tSNE"}

\item{clusterMethod}{indicates the method to do clustering: "dbscan" or "PAM"}
}
\value{
score dataframe with the following elements: SR, DPT, potency_states
}
\description{
The function apply Landscent package to a gene by cell matric to compute signal entropy, potency states and diffusion pseudo time
}
\author{
Noemi Di Nanni
}
