% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscent.R
\name{landscent}
\alias{landscent}
\title{landscent}
\usage{
landscent(
  genes_by_cells_matrix,
  ppi = NULL,
  reduceMethod = "PCA",
  clusterMethod = "dbscan",
  mc.cores = 2
)
}
\arguments{
\item{genes_by_cells_matrix}{real data}

\item{ppi}{The adjacency matrix of a user-given PPI network with rownames and colnames labeling genes (same gene identifier as in genes_by_cells_matrix)}

\item{reduceMethod}{indicates the method to do dimension reduction: "PCA" or "tSNE"}

\item{clusterMethod}{indicates the method to do clustering: "dbscan" or "PAM"}
}
\value{
landscent_list list with the following elements: SR, DPT, potency_states, complete_output
}
\description{
landscent
}
