% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_matrix.R
\name{merge_matrix}
\alias{merge_matrix}
\title{merge_matrix}
\usage{
merge_matrix(feature_list = NULL)
}
\arguments{
\item{signatures_by_cells}{matrix, signatures-by-cells input}

\item{expr_score}{matrix, expression scores (output of exp_Rate)}

\item{output_landscent}{dataframe, df with the following elements: SR, DPT, potency_states (output of landSCENT)}
}
\value{
matrix_merged seurat object, object with the merged and scaled matrix
}
\description{
Create and scale a matrix with different features (row) for each cell (column), like signature/landscend scores
}
\author{
Noemi Di Nanni
}
