% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_cluster.R
\name{meta_cluster}
\alias{meta_cluster}
\title{Returns meta-clusters obtained by using hierarchical clustering}
\usage{
meta_cluster(overlap_mat, n_step = 11, max_clust = 10, do_plot = FALSE)
}
\arguments{
\item{overlap_mat}{Overlap matrix as obtained form overlap_matrix() function}

\item{n_step}{number used to divide hclust dendrogram height to obtain cut values to calculate silhouette}

\item{max_clust}{maximum number of clusters}

\item{do_plot}{whether to do the silohouette plot or not}
}
\description{
Returns meta-clusters obtained by using hierarchical clustering
}
