% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_matrix.R
\name{overlap_matrix}
\alias{overlap_matrix}
\title{Calculate the overlap coefficient between n sets}
\usage{
overlap_matrix(scMuffinList = NULL)
}
\arguments{
\item{x}{list of item sets}
}
\description{
Calculate the overlap coefficient between n sets
}
