% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions_to_list.R
\name{partitions_to_list}
\alias{partitions_to_list}
\title{clusterings_to_list}
\usage{
partitions_to_list(partitions)
}
\arguments{
\item{partitions}{partitions object}
}
\description{
clusterings_to_list
}
