% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_dataset_comparison.R
\name{plot_heatmap_dataset_comparison}
\alias{plot_heatmap_dataset_comparison}
\title{plot_heatmap_dataset_comparison}
\usage{
plot_heatmap_dataset_comparison(
  dataset_cmp_list = NULL,
  type = "score",
  show.gs.source = FALSE,
  outfile = NULL,
  width = 200,
  height = 200,
  units = "mm",
  res = 300,
  ...
)
}
\arguments{
\item{type}{"score" or "significance" for, respectively, cluster median marker set score or -log10(FDRq) of GSEA.}

\item{outfile}{out file name}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image res}

\item{...}{arguments passed to ComplexHeatmap::Heatmap}
}
\description{
plot_heatmap_dataset_comparison
}
