% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_features_by_clusters.R
\name{plot_heatmap_features_by_clusters}
\alias{plot_heatmap_features_by_clusters}
\title{plot_heatmap_features_by_clusters}
\usage{
plot_heatmap_features_by_clusters(
  features_by_clusters,
  significance_matrix = NULL,
  sig_threshold = 0.05,
  ntop = 10,
  onlyUp = TRUE,
  out_dir = "./",
  remove_null_features = FALSE,
  ...
)
}
\arguments{
\item{features_by_clusters}{list resulting from calculate_signatures_clusters()}

\item{significance_matrix}{significance matrix to add asterisks over the heatmap}

\item{sig_threshold}{significance threshold}

\item{ntop}{number of top features considered for each cluster}

\item{onlyUp}{top features are considered only on the basis of their positive deviation from null distribution (up-regulation)}

\item{out_dir}{output directory}

\item{remove_null_features}{whether to remove null features}

\item{...}{further arguments to ComplexHeatmap::Heatmap}
}
\description{
plot_heatmap_features_by_clusters
}
