% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_features_by_clusters.R
\name{plot_heatmap_features_by_clusters}
\alias{plot_heatmap_features_by_clusters}
\title{Plot an heatmap of feature values in cell clusters}
\usage{
plot_heatmap_features_by_clusters(
  scMuffinList = NULL,
  feature_source = NULL,
  partition_id = NULL,
  significance_matrix = NULL,
  sig_threshold = 0.05,
  file = NULL,
  remove_null_features = FALSE,
  width = 180,
  height = 180,
  units = "mm",
  res = 300,
  scale = TRUE,
  pal = NULL,
  ...
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{feature_source}{It can be a "mean", "gss" or a numeric matrix (clusters-by-features). If "mean", the data.frame with average feature values among clusters will be used (default); if "gene_set_scoring", the average gene set values among clusters will be used.}

\item{partition_id}{identifier of the partition to be considered}

\item{significance_matrix}{optional significance matrix (clusters-by-features) of the same size of the data specified by means of feature_source}

\item{sig_threshold}{significance threshold}

\item{file}{output file}

\item{remove_null_features}{whether to remove null features}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{scale}{whether to scale the features}

\item{pal}{color palette. Default to rev(pals::brewer.rdylbu(10)) (negative values) or pals::brewer.ylorrd(5)) (positive values)}

\item{...}{further arguments to ComplexHeatmap::Heatmap}
}
\description{
Plot an heatmap of feature values in cell clusters
}
