% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile_CNV.R
\name{plot_profile_CNV}
\alias{plot_profile_CNV}
\title{CNV Profile}
\usage{
plot_profile_CNV(
  scMuffinList = NULL,
  cluster = 0,
  z.score = TRUE,
  file = NULL,
  width = 300,
  height = 90,
  units = "mm",
  res = 300,
  cex.points = 0.7,
  cex.lab = 0.7,
  cex.axis = 0.7,
  cex.main = 0.7
)
}
\arguments{
\item{scMuffinList}{scMuffinList object}

\item{cluster}{cluster to be plotted}

\item{z.score}{whether to plot cluster median z-scores of CNV profile (TRUE) or cluster median CNV profile (FALSE).}

\item{file}{File name to save the figure as png file}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{cex.points}{cex of points}

\item{cex.lab}{cex of labs}

\item{cex.axis}{cex of axis}

\item{cex.main}{cex of main}
}
\description{
Plot an heatmap of the CNV.
}
\details{
CNV Profile of every cluster
}
