% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_colored_features.R
\name{plot_umap_colored_features}
\alias{plot_umap_colored_features}
\title{UMAP Visualization colored by features}
\usage{
plot_umap_colored_features(
  Seu_obj = NULL,
  scMuffinList = NULL,
  feature_name = NULL,
  scale_feature = FALSE,
  adj_outliers = FALSE,
  min_cells = 10,
  out_dir = "./",
  width = 180,
  height = 180,
  units = "mm",
  res = 300,
  ...
)
}
\arguments{
\item{Seu_obj}{Seurat object with UMAP data}

\item{scMuffinList}{scMuffinList object}

\item{feature_name}{feature name}

\item{scale_feature}{whether to scale feature values}

\item{adj_outliers}{logical, whether to adjust the feature scores, removing outliers}

\item{min_cells}{minimum number of cells in which the feature must have a non-zero value}

\item{out_dir}{output directory}

\item{width}{image width}

\item{height}{image height}

\item{units}{image units}

\item{res}{image resolution}

\item{...}{further arguments to plot_umap}
}
\description{
Generate UMAP visualizations for all the columns in the "summary" element of a feature available in scMuffinList. For every column a png file is saved in out_dir. It requires a Seurat object with UMAP information.
}
