% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_colored_features.R
\name{plot_umap_colored_features}
\alias{plot_umap_colored_features}
\title{plot_umap_colored_features}
\usage{
plot_umap_colored_features(seurat_object, features_by_cells)
}
\arguments{
\item{seurat_object}{seurat object, object with saved dimension reduction components calculate on genes by cells matrix}

\item{features_by_cells}{matrix, features by cells matrix}

\item{n_colors}{numeric, number of colors to be used}
}
\description{
Generate a UMAP visualization colored by feature distributions
}
