% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_colored_features.R
\name{plot_umap_colored_features}
\alias{plot_umap_colored_features}
\title{plot_umap_colored_features}
\usage{
plot_umap_colored_features(
  Seu_obj,
  features,
  dir = "./",
  scale_feature = TRUE,
  feature_breaks = NULL,
  adj_outliers = FALSE,
  neg_eps = -0.1,
  pos_eps = 0.1,
  min_cells = 10,
  ...
)
}
\arguments{
\item{Seu_obj}{seurat object, object with saved dimension reduction components calculate on genes by cells matrix}

\item{features}{matrix, features by cells matrix}

\item{dir}{output directory}

\item{scale_feature}{whether to scale features or not}

\item{feature_breaks}{breaks for feature value distribution}

\item{adj_outliers}{logical, whether to adjust the feature scores, removing outliers}

\item{neg_eps}{small negative value to define the color for null values around zero, default to -0.1}

\item{pos_eps}{small positive value to define the color for null values around zero, default to =0.1}

\item{min_cells}{min number of cells in which the feature must have a non-zero value}

\item{...}{further arguments to plot_umap}
}
\description{
Generate a UMAP visualization colored by feature distributions
}
