% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_colored_features.R
\name{plot_umap_colored_features}
\alias{plot_umap_colored_features}
\title{plot_umap_colored_features}
\usage{
plot_umap_colored_features(
  Seu_obj = NULL,
  scMuffinList = NULL,
  feature_id = NULL,
  scale_feature = TRUE,
  n_intervals = 10,
  adj_outliers = FALSE,
  min_cells = 10,
  ...
)
}
\arguments{
\item{Seu_obj}{seurat object, object with saved dimension reduction components calculate on genes by cells matrix}

\item{scMuffinList}{scMuffinList object}

\item{feature_id}{feature id}

\item{scale_feature}{whether to scale features or not}

\item{n_intervals}{number of intervals for the palette}

\item{adj_outliers}{logical, whether to adjust the feature scores, removing outliers}

\item{min_cells}{min number of cells in which the feature must have a non-zero value}

\item{...}{further arguments to plot_umap}

\item{feature_breaks}{breaks for feature value distribution}
}
\description{
Generate a UMAP visualization colored by feature distributions
}
\details{
Generate a UMAP visualization colored by feature distributions
}
