% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_object_for_CNV.R
\name{preprocess_object_for_CNV}
\alias{preprocess_object_for_CNV}
\title{preprocess_object_for_CNV}
\usage{
preprocess_object_for_cnv(input_object)
}
\arguments{
\item{input_object}{Seurat object.}
}
\value{
list of dataframes, where every dataframe is a chromosome
}
\description{
Preprocessing pipeline to obtain a list of ordered dataframes (one for each chromosome).
}
\details{
Preprocessing consist of annotation, removing duplicates and NA values.
}
\author{
Valentina Nale
}
