% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_object_for_CNV.R
\name{preprocess_object_for_CNV}
\alias{preprocess_object_for_CNV}
\title{preprocess_object_for_CNV}
\usage{
preprocess_object_for_CNV(genes_by_cells)
}
\arguments{
\item{genes_by_cells}{genes-by-cells input matrix}
}
\value{
list of genomically-ordered chromosomes
}
\description{
Preprocessing function to obtain a genomically-ordered list of chromosomes.
}
\details{
The preliminary step consist of annotation, duplicates and NA values removal. 
Then, the matrix is splitted as a list of dataframe, where every dataframe is a chromosome.
Chromosomes are ordered from 1 to 22 + X +Y, and then re-ordered by start position.
}
\author{
Valentina Nale
}
