% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_create_null.R
\name{sc_create_null}
\alias{sc_create_null}
\title{sc_create_null}
\usage{
sc_create_null(seurat_data, bins, gene_set, k = 100)
}
\arguments{
\item{seurat_data}{seurat data object}

\item{bins}{bins}

\item{gene_set}{genes}

\item{k}{number of permutations}
}
\description{
sc_create_null
}
\author{
Ettore Mosca
}
