% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_create_null.R
\name{sc_create_null}
\alias{sc_create_null}
\title{sc_create_null}
\usage{
sc_create_null(genes_by_cells, bins, gene_set, k = 100)
}
\arguments{
\item{genes_by_cells}{genes-by-cells matrix}

\item{bins}{bins of genes}

\item{gene_set}{a vector of genes}

\item{k}{number of permutations}
}
\description{
sc_create_null
}
\author{
Ettore Mosca
}
