% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_data_bin.R
\name{sc_data_bin}
\alias{sc_data_bin}
\title{Split the genes into bins by their expression}
\usage{
sc_data_bin(genes_by_cells = NULL, nbins = 25, na.rm = FALSE)
}
\arguments{
\item{genes_by_cells}{genes-by-cells matrix}

\item{nbins}{number of bins}

\item{na.rm}{if TRUE, cells with null (0) values are not considered in the calculation of the gene expression level.}
}
\value{
vector with bin identifier for each gene
}
\description{
Split the input data matrix by rows into nbins
}
