% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_data_bin.R
\name{sc_data_bin}
\alias{sc_data_bin}
\title{sc_data_bin}
\usage{
sc_data_bin(seurat_data, nbins = 25, use.log = TRUE)
}
\arguments{
\item{seurat_data}{seurat data object}

\item{nbins}{number of bins}

\item{use.log}{use logarithm}
}
\value{
vector with bin id for each row
}
\description{
Split the input data matrix by rows into nbins
}
