% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMuffin.R
\name{scMuffin}
\alias{scMuffin}
\title{scMuffin: main function that runs the whole pipeline}
\usage{
scMuffin(
  genes_by_cells,
  custom_signatures = NULL,
  analyses = c("signatures", "landscent", "cnv", "expr_rate"),
  mc.cores = 2
)
}
\arguments{
\item{genes_by_cells}{seurat_object_data, genes-by-cells input matrix}

\item{custom_signatures}{list of gene sets}
}
\description{
scMuffin: main function that runs the whole pipeline
}
