% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_cnv.R
\name{score_cnv}
\alias{score_cnv}
\title{Score CNV}
\usage{
score_cnv(cnv_res, plot_score = FALSE)
}
\arguments{
\item{cnv_res}{Object resulting from calculate_cnv function}

\item{plot_score}{NULL. Set as TRUE it return a plot.}
}
\description{
Score CNV
}
