% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tissues.R
\name{show_tissues}
\alias{show_tissues}
\title{Show the tissues available by means of prepare_gsl}
\usage{
show_tissues(gsc_source = c("SIG_PNDB", "SIG_CM_normal", "SIG_CM_cancer"))
}
\arguments{
\item{gsc_source}{the source database}
}
\description{
Show the tissues available by means of prepare_gsl
}
